Control.Auto

Types

Output

data Output m a b

The output of a stepAuto. Contains the "result" value of the stepping (outRes), and the "next Auto", outAuto.

An Auto m a b will produce an Output m a b; when "stepped" with an a, the "result" (outRes) is a b.

Really, you can just think of this as a fancy tuple.

Constructors

Output

Fields

outRes :: b

Result value of a step

outAuto :: Auto m a b

The next Auto

type Output' = Output Identity

Special case of Output where the underlying Monad of outAuto is Identity.

onOutput

Arguments

:: (b -> b')
function over the result

-> (Auto m a b -> Auto m a' b')
function over the resulting Auto

-> Output m a b

-> Output m a' b'

Map two functions onto both fields of an Output.

If you want to map an a -> b onto both fields (the result and the result of the next Auto), you can use the Functor instance instead.

Auto

data Auto m a b

The Auto type. Basically represents a function containing its own localized internal state. If you have an Auto a b, you can "step" it with stepAuto and an a, to get a b and a "next Auto". The a is the input, and the b is the output, and the next Auto is the Auto with updated internal state.

The "stepping" process can be monadic:

stepAuto :: Auto m a b -> a -> m (Output m a b)

So you can have access to, say, a shared environment using Reader or something like that.

Auto is mostly useful because of its Functor, Applicative, Category, and Arrow (and Arrow-related) instances. These allow you to modify, combine, chain, and side-chain Autos in expressive way, allowing you to build up complex ones from combinations of simple, primitive ones.

TODO: see tutorial

The Auto also contains information on its own serialization, so you can serialize and re-load the internal state without actually accessing it.

type Auto' = Auto Identity

Special case of Auto where the underlying Monad is Identity.

Misc

data Blip a

A type representing a "discrete" sort of event-like thing. It's supposed to represent something that happens alone, and one at a time, as opposed to things that are "on" or "off" for large intervals at a time.

It's here mainly because it's a pretty useful abstraction in the context of the many combinators found in various modules of this library. If you think of an Auto m a (Blip b) as a "Blip stream", then there are various combinators and functions that are specifically designed to manipulate "Blip streams".

For the purposes of the semantics of what Blip is supposed to represent, its constructors are hidden. (Almost) all of the various Blip combinators (and its very useful Functor instance) "preserve Blipness" --- one-at-a-time occurrences remain one-at-a-time under all of these combinators, and you should have enough so that direct access to the constructor is not needed.

If you are creating a framework, library, or backend, you might want to manually create Blip stream-producing Autos for your users to access. In this case, you can import the constructors and useful internal (and, of course, semantically unsafe) functions from Control.Auto.Blip.Internal.

type Interval m a b = Auto m a (Maybe b)

An Interval is "just" a type alias for Auto m a (Maybe b). If you ended up here with a link...no worries! If you see Interval m a b, just think Auto m a (Maybe b)!

type Interval' a b = Auto' a (Maybe b)

Interval, specialized with Identity as its underlying Monad. (Analogous to Auto' for Auto)

Running

stepAuto

Arguments

:: Monad m

=> Auto m a b
the Auto to step

-> a
the input

-> m (Output m a b)
the output, and the updated Auto'.

Runs the Auto through one step.

Remember that at every step for an Auto m a b, you provide an a input and receive a b output with an "updated"/"next" Auto.

>>> let a = mkAccum (+) 0 :: Auto Identity Int Int
            -- an Auto that sums all of its input.
>>> let Identity (Output y a') = stepAuto a 3
>>> y      -- the result
3 :: Int
>>> :t a'   -- the updated 'Auto'
a' :: Auto Identity Int Int

(Identity, from Data.Functor.Identity, is the "dumb Functor": data Identity a = Identity a)

If you think of an Auto m a b as a "stateful function" of type a -> m b, then stepAuto lets you "run" it.

stepAuto'

Arguments

:: Auto' a b
the Auto' to step

-> a
the input

-> Output' a b
the output, and the updated Auto' stepAuto' a = runIdentity . stepAuto a

stepAuto, but for an Auto' --- the underlying Monad is Identity. Returns the output stripped of Identity.

If you think of an Auto' a b as a "stateful function" of type a -> b, then stepAuto' lets you "run" it.

evalAuto :: Monad m => Auto m a b -> a -> m b

Like stepAuto, but drops the "next Auto" and just gives the result.

evalAuto' :: Auto' a b -> a -> b

Like stepAuto', but drops the "next Auto'" and just gives the result. evalAuto for Auto'.

streamAuto

Arguments

:: Monad m

=> Auto m a b
Auto to stream

-> [a]
input stream

-> m [b]
output stream

Stream an Auto over a list, returning the list of results. Does this "lazily" (over the Monad), so with most Monads, this should work fine with infinite lists.

Note that, conceptually, this turns an Auto m a b into an [a] -> m [b].

streamAuto'

Arguments

:: Auto' a b
Auto' to stream

-> [a]
input stream

-> [b]
output stream

Stream an Auto' over a list, returning the list of results. Does this lazily, so this should work fine with (and is actually somewhat designed for) infinite lists.

Note that conceptually this turns an Auto' a b into an [a] -> [b]

stepAutoN

Arguments

:: Monad m

=> Int
number of times to step the Auto

-> Auto m a b
the Auto to run

-> a
the repeated input

-> m ([b], Auto m a b)
list of outputs and the updated Auto

Repeatedly steps an Auto with the same input a given number of times.

stepAutoN n a0 x = overList a0 (replicate n x)

>>> let a          = iterator (*2) 1
>>> let (ys, a')   = runIdentity (stepAutoN 8 a ())
>>> ys
[1, 2, 4, 8, 16, 32, 64, 128]
>>> let Output y _ = runIdentity (stepAuto a ())
>>> y
256

stepAutoN'

Arguments

:: Int
number of times to step the Auto'

-> Auto' a b
the Auto' to run

-> a
the repeated input

-> ([b], Auto' a b)
list of outputs and the updated Auto'

Like stepAutoN, but with an Auto' (the underlying Monad is Identity)

>>> let a          = iterator (*2) 1
>>> let (ys, a')   = stepAutoN 8 a ()
>>> ys
[1, 2, 4, 8, 16, 32, 64, 128]
>>> let Output y _ = stepAuto a ()
>>> y
256

Serializing

See the header of the "serializing" section of Control.Auto.Core for more detail on how these work.

encodeAuto :: Auto m a b -> ByteString

Encode an Auto and its internal state into a ByteString.

decodeAuto :: Auto m a b -> ByteString -> Either String (Auto m a b)

Resume an Auto from its ByteString serialization, giving a Left if the deserialization is not possible.

readAuto

Arguments

:: FilePath
filepath to read from

-> Auto m a b
Auto to resume

-> IO (Either String (Auto m a b))

Give a FilePath and an Auto, and readAuto will attempt to resume the saved state of the Auto from disk, reading from the given FilePath. Will return Left upon a decoding error, with the error, and Right if the decoding is succesful.

writeAuto

Arguments

:: FilePath
filepath to write to

-> Auto m a b
Auto to serialize

-> IO ()

Given a FilePath and an Auto, serialize and freeze the state of the Auto as binary to that FilePath.

Strictness

forcer :: NFData a => Auto m a a

A special Auto that acts like the id Auto, but forces results as they come through to be fully evaluated, when composed with other Autos.

seqer :: Auto m a a

A special Auto that acts like the id Auto, but forces results as they come through to be evaluated to Weak Head Normal Form, with seq, when composed with other Autos.

Fixed points

lastVal

Arguments

:: Serialize a

=> a
initial value

-> Auto m a a

An Auto that returns the last value received by it. Given an "initial value" to output first.

This is (potentially) a very dangerous Auto, because its usage and its very existence opens the door to breaking denotative/declarative style and devolving into imperative style coding. However, when used where it is supposed to be used, it is more or less invaluable, and will be an essential part of many programs.

Its main usage is for dealing with recursive bindings. If you ever are laying out recursive bindings in a high-level/denotative way, you need to have at least one value be able to have a "initial output" without depending on anything else. lastVal and delay allow you to do this.

See the recursive example for more information on the appropriate usage of lastVal and delay.

lastVal_

Arguments

:: a
initial value

-> Auto m a a

The non-resuming/non-serializing version of lastVal.

Auto constructors

arrM

Arguments

:: (a -> m b)
monadic function

-> Auto m a b

Applies the given "monadic function" (function returning a monadic action) to every incoming item; the result is the result of executing the action returned.

Note that this essentially lifts a "Kleisli arrow"; it's like arr, but for "monadic functions" instead of normal functions:

arr  :: (a -> b)   -> Auto m a b
arrM :: (a -> m b) -> Auto m a b

arrM f . arrM g == arrM (f <=< g)

from Accumulators

Result-first

mkAccum

Arguments

:: Serialize b

=> (b -> a -> b)
accumulating function

-> b
initial accumulator

-> Auto m a b

Construct an Auto from a "folding" function: b -> a -> b yields an Auto m a b. Basically acts like a foldl or a scanl. There is an internal accumulator that is "updated" with an a at every step. Must be given an initial accumulator.

Example: an Auto that sums up all of its input.

>>> let summer = mkAccum (+) 0
>>> let Output sum1 summer' = stepAuto summer 3
>>> sum1
3
>>> let Output sum2 _       = stepAuto summer'' 10
>>> sum2
13

If your accumulator b does not have a Serialize instance, then you should either write a meaningful one, or throw away serializability and use mkAccum_.

mkAccum_

Arguments

:: (b -> a -> b)
accumulating function

-> b
intial accumulator

-> Auto m a b

A version of 'mkAccum_, where the internal accumulator isn't serialized. It can be "saved" and "loaded", but the state is lost in the process.

Useful if your accumulator b cannot have a meaningful Serialize instance.

mkAccumM

Arguments

:: (Serialize b, Monad m)

=> (b -> a -> m b)
(monadic) accumulating function

-> b
initial accumulator

-> Auto m a b

Construct an Auto from a "monadic" "folding" function: b -> a -> m b yields an Auto m a b. Basically acts like a foldM or scanM (if it existed). here is an internal accumulator that is "updated" with an input a with the result of the executed m b at every step. Must be given an initial accumulator.

If your accumulator b does not have a Serialize instance, then you should either write a meaningful one, or throw away serializability and use mkAccumM_.

mkAccumM_

Arguments

:: Monad m

=> (b -> a -> m b)
(monadic) accumulating function

-> b
initial accumulator

-> Auto m a b

A version of 'mkAccumM_, where the internal accumulator isn't serialized. It can be "saved" and "loaded", but the state is lost in the process.

Useful if your accumulator b cannot have a meaningful Serialize instance.

Initial accumulator-first

mkAccumD

Arguments

:: Serialize b

=> (b -> a -> b)
accumulating function

-> b
initial accumulator

-> Auto m a b

A "delayed" version of mkAccum, where the first output is actually the initial state of the accumulator. Useful in recursive bindings.

mkAccumD_

Arguments

:: (b -> a -> b)
accumulating function

-> b
intial accumulator

-> Auto m a b

The non-resuming/non-serializing version of mkAccumD.

mkAccumMD

Arguments

:: (Serialize b, Monad m)

=> (b -> a -> m b)
(monadic) accumulating function

-> b
initial accumulator

-> Auto m a b

A "delayed" version of mkAccumM, where the first output is actually the initial state of the accumulator. Useful in recursive bindings.

mkAccumMD_

Arguments

:: Monad m

=> (b -> a -> m b)
(monadic) accumulating function

-> b
initial accumulator

-> Auto m a b

The non-resuming/non-serializing version of mkAccumMD.

from State transformers

mkState

Arguments

:: Serialize s

=> (a -> s -> (b, s))
state transformer

-> s
intial state

-> Auto m a b

Construct an Auto from a state transformer: an a -> s -> (b, s) gives you an Auto m a b, for any Monad m. At every step, it takes in the a input, runs the function with the stored internal state, returns the b result, and now contains the new resulting state. You have to intialize it with an initial state, of course.

Try not to use this if it's ever avoidable, unless you're a framework developer or something. Try make something by combining/composing the various Auto combinators.

This version is a wrapper around mkAuto, that keeps track of the serialization and re-loading of the internal state for you, so you don't have to deal with it explicitly.

If your state s does not have a Serialize instance, then you should either write a meaningful one, provide the serialization methods manually with mkState', or throw away serializability and use mkState_.

mkStateM

Arguments

:: Serialize s

=> (a -> s -> m (b, s))
(monadic) state transformer

-> s
initial state

-> Auto m a b

Construct an Auto from a "monadic" state transformer: a -> s -> m (b, s) gives you an Auto m a b. At every step, it takes in the a input, runs the function with the stored internal state and "executes" the m (b, s) to get the b output, and stores the s as the new, updated state. Must be initialized with an initial state.

Try not to use this if it's ever avoidable, unless you're a framework developer or something. Try make something by combining/composing the various Auto combinators.

This version is a wrapper around mkAuto, that keeps track of the serialization and re-loading of the internal state for you, so you don't have to deal with it explicitly.

If your state s does not have a Serialize instance, then you should either write a meaningful one, provide the serialization methods manually with mkStateM', or throw away serializability and use mkStateM_.

mkState_

Arguments

:: (a -> s -> (b, s))
state transformer

-> s
initial state

-> Auto m a b

A version of mkState, where the internal state isn't serialized. It can be "saved" and "loaded", but the state is lost in the process.

Useful if your state s cannot have a meaningful Serialize instance.

mkStateM_

Arguments

:: (a -> s -> m (b, s))
(monadic) state transformer

-> s
initial state

-> Auto m a b

A version of mkStateM, where the internal state isn't serialized. It can be "saved" and "loaded", but the state is lost in the process.

Useful if your state s cannot have a meaningful Serialize instance.

Generators

Effects

effect

Arguments

:: m b
monadic action to contually execute.

-> Auto m a b

To get every output, executes the monadic action and returns the result as the output. Always ignores input.

This is basically like an "effectful" pure:

pure   :: b   -> Auto m a b
effect :: m b -> Auto m a b

The output of pure is always the same, and the output of effect is always the result of the same monadic action. Both ignore their inputs.

Fun times when the underling Monad is, for instance, Reader.

>>> let a = effect ask    :: Auto (Reader b) a b
>>> let r = evalAuto a () :: Reader b b
>>> runReader r "hello"
"hello"
>>> runReader r 100
100

exec

Arguments

:: Monad m

=> m b
monadic action to contually execute.

-> Auto m a a

Acts like id, in that the output stream is identical to the input stream. However, at each retrieval of output, executes the given monadic action and ignores the result.

Iterators

iterator

Arguments

:: Serialize b

=> (b -> b)
iterating function

-> b
starting value and initial output

-> Auto m a b

Analogous to iterate from Prelude. Keeps accumulator value and continually applies the function to the value at every step, outputting the result.

The first result is the initial accumulator value.

>>> let (y, _) = stepAutoN' 10 (iterator (*2) 1) ()
>>> y
[1, 2, 4, 8, 16, 32, 64, 128, 256, 512]

iterator_

Arguments

:: (b -> b)
iterating function

-> b
starting value and initial output

-> Auto m a b

The non-resuming/non-serializing version of iterator.

iteratorM

Arguments

:: (Serialize b, Monad m)

=> (b -> m b)
(monadic) iterating function

-> b
starting value and initial output

-> Auto m a b

Like iterator, but with a monadic function.

iteratorM_

Arguments

:: Monad m

=> (b -> m b)
(monadic) iterating function

-> b
starting value and initial output

-> Auto m a b

The non-resuming/non-serializing version of iteratorM.

Common Autos and combinators

Processes

sumFrom

Arguments

:: (Serialize a, Num a)

=> a
initial count

-> Auto m a a

The stream of outputs is the cumulative/running sum of the inputs so far, starting with an initial count.

The first output takes into account the first input. See sumFromD for a version where the first output is the initial count itself.

sumFrom x0 = mkAccum (+) x0

productFrom

Arguments

:: (Serialize a, Num a)

=> a
initial product

-> Auto m a a

The output is the running/cumulative product of all of the inputs so far, starting from an initial product.

productFrom x0 = mkAccum (*) x0

mappender :: (Serialize a, Monoid a) => Auto m a a

The output is the running/cumulative mconcat of all of the input seen so far, starting with mempty.

>>> streamauto' mappender . map Last $ [Just 4, Nothing, Just 2, Just 3]
[Last (Just 4), Last (Just 4), Last (Just 2), Last (Just 3)]
>>> streamAuto' mappender ["hello","world","good","bye"]
["hello","helloworld","helloworldgood","helloworldgoodbye"]

mappender = mkAccum mappend mempty

mappendFrom

Arguments

:: (Serialize a, Semigroup a)

=> a
initial value

-> Auto m a a

The output is the running <>-sum (mappend for Semigroup) of all of the input values so far, starting with a given starting value. Basically like mappender, but with a starting value.

>>> streamAuto' (mappendFrom (Max 0)) [Max 4, Max (-2), Max 3, Max 10]
[Max 4, Max 4, Max 4, Max 10]

mappendFrom m0 = mkAccum (<>) m0

Switches

(-->) infixr 1

Arguments

:: Monad m

=> Interval m a b
initial behavior

-> Auto m a b
final behavior, when the initial behavior turns off.

-> Auto m a b

"One after the other". Behave like the first Auto (and run only its effects) as long as it is "on" (Just). As soon as it is Nothing, begin behaving like the second Auto forevermore, and running the second Auto's effects, if any. Works well if the Autos follow interval semantics from Control.Auto.Interval.

>>> let a1 = (onFor 2 . pure "hello") --> pure "world"
>>> let Output res1 _ = stepAutoN' 5 a1 ()
>>> res1
["hello", "hello", "world", "world", "world"]

(Recall that pure "hello" is the Auto that ignores its input and constantly outputs "hello", and onFor 2 lets its input pass "on" (Just) for two steps, then is "off" (Nothing) ever after.)

Association works in a way that you can "chain" -->s, as long as you have a non-Maybe / non-Interval Auto at the end.

>>> let a2 = onFor 2 . pure "hello"
         --> onFor 2 . pure "world"
         --> pure "goodbye!"
>>> let Output res2 _ = stepAutoN' 6 a2 ()
>>> res2
["hello", "hello", "world", "world", "goodbye!", "goodbye!"]

a --> b --> c associates as a --> (b --> c)

This is pretty invaluable for having Autos "step" through a series of different Autos, progressing their state from one stage to the next. Autos can control when they want to be "moved on" from by turning "off" (outputting Nothing).

Note that recursive bindings work just fine, so:

>>> let a3 = onFor 2 . pure "hello"
         --> onFor 2 . pure "world"
         --> a3
>>> let (res3, _) = stepAutoN' 8 a3 ()
>>> res3
["hello", "hello", "world", "world", "hello", "hello", "world", "world"]

the above represents an infinite loop between outputting "hello" and outputting "world".

By the way, it might be worth contrasting this with <|!> and <|?> from Control.Auto.Interval, which have the same type signatures. Those alternative-y operators always feed the input to both sides, run both sides, and output the first Just. With <|!>, you can "switch back and forth" to the first Auto as soon as the first Auto is "on" (Just) again.

-->, in contrast, runs only the first Auto until it is off (Nothing)...then runs only the second Auto. This transition is one-way, as well.

Running

interactAuto

Arguments

:: Interval' String String
Interval' to run interactively

-> IO (Interval' String String)
final Interval' after it all

Run an Auto' "interactively". Every step grab a string from stdin, and feed it to the Auto. If the Auto pops out Nothing, end the session; if it outputs a Just, print it out to stdout and repeat all over again.

If your Auto outputs something other than a String, you can use fmap to transform the output into a String en-route (like fmap show).

If your Auto takes in something other than a String, you can lmap a function to convert the input String to whatever intput your Auto expects.

You can use duringRead or bindRead if you have an Auto' or Interval' that takes something readable, to chug along until you find something non-readable.

Outputs the final Auto' when the interaction terminates.

interactRS

Arguments

:: (Read a, Show b)

=> Interval' a b
Interval' to run interactively

-> IO (Interval' String String)
final Interval' after it all

Like interact, but instead of taking Interval' String String, takes any Interval' a b as long as a is Read and b is Show.

Will "stop" if either (1) the input is not read-able or (2) the Interval' turns off.

Outputs the final Auto' when the interaction terminates.

Re-exports

module Control.Category

module Control.Applicative

module Control.Arrow

module Data.Semigroup