Control.Auto.Blip

Description

This module provides tools for generating and manipulating Blip streams. The Blip abstraction is not fundamental to Auto, but rather, like interval, is a very useful semantic tool for the denotation of many programs, games, simulations, and computations in general that you are likely to write with this library.

Blip

A Blip a, in the context of Auto and Auto compositions, represents a "stream of Blips". This is a stream that may "emit" a value every once-in-a-while, "discretely", or in "isolated" occurrences.

By this, we mean that every emitted value is emitted "alone", and not as a part of continuous on/off chunks.

Example situations

A good example would be, say, a blip stream that emits every time a user/player sends a certain type of command. Or a blip stream that emits every time a slowly-moving value crosses over from positive to negative.

A bad example would be a blip stream that emits when a player doesn't send a certain less-common type of command. Or a blip stream that emits whenever a slowly-moving value is positive or negative.

Contrast with Intervals

Blip streams are contrasted with another semantic tool: stream _intervals_, manipulated with Control.Auto.Interval. Intervals are adjacentcontiguous "chunks" of onoff behavior, and are on or off for contiguous "chunks" at a time. So when deciding whether or not you want to use the semantics of Blip or the semantics of Interval, consider: is this behavior going to be "on/off" for chunks at a time (such as an interval that is on whenever a slowly-moving value is positive)? Or is it something that is usually "not on", but makes separate, isolated, "blips" --- each emitted value alone and (semantically) isolated from the rest.

Motivations

The main motivations of the semantic concept of Blips (and why they even exist in the first place) is probably for how well they integrate (with interval semantics) with the various powerful switching combinators from Control.Auto.Switch. Many of the combinators in that module are designed so that switches can be "triggered" by Blips.

Blips have many usages, as will be explained later. You'll also find that Blip streams work well with their cousins, interval streams. But perhaps the use case that stands out above all (and is alone enough to motivate their existence) is in switching.

"Blip semantics"

We say that a Blip stream has "Blip semantics" when it is used in a way that its emitted values are "isolated", "alone", "discrete", in this way. When it is not, we say that the stream "breaks" Blip semantics.

Note that this can't really be enforced by the types, so it's up to you as a library developer or application developer to take care that the blip streams you make all conform to Blip semantics.

Also note that in many of these cases, "Blip semantics" depends on how the Autos are composed, and what they are composed to. If the value in question is "almost always" positive and only negative at isolated points in time, then such a "blip stream that emits whenever the value is negative" has proper Blip semantics. If the value in question is slowly-moving and meandering, and might spend a lot of time negative at a time, then the same blip stream would not preserve Blip semantics.

Why semantics are important

Why should you care? I can't tell you want to do, right?

Well, for the same reason that typeclasses like Eq, Functor, and Monad have laws. Yeah, you can make any instance you want that satisfies the types. But almost all of the usefulness of those typeclasses comes from our ability to "reason" about the behavior of their instances, and to be able to develop an intuition about their usage. We would be surprised if we had an Eq instance where x == x and x /= x are both true...and it would completely break down any attempt at understanding what Eq code "means".

You can think of "Blip semantics" as being the "laws" of Blip streams. If we assume that things follow blip semantics properly, then we can reason about them in a unified and useful way. If we can trust that Blips actually behave "like blips", then Blip becomes an extremely useful tool for denoting certain behaviors and programs.

If we can't...then it becomes a lot less useful :)

In particular, one big use case for Blip streams (the switching mechanisms Control.Auto.Switch) all only "work well" when your Blip streams follow proper semantics.

Combinators preserve semantics

Most of the combinators in this module try their best to preserve Blip semantics. That is, you can't use them in a way that will produce a non-semantic-abiding Blip stream. You can "trust" them, and if you use only safe combinators, you don't ever have to worry. Well. That much, at least.

There are a few notable exceptions:

every, eachAt, eachAt_, when you pass in an interval of 1.

onChange, when the input value isn't ever expected to stay the same between steps.

emitOn, emitJusts, onJusts, in the cases mentioned in the documentation for emitOn.

Practical examples

There are many practical examples of using Blips in the various examples in auto-examples, especially from chatbot. There, Blip streams are used in many situations, primarily streams for players sending certain commands. It's also used in hangman, to signify player events such as victory, good/bad guesses, etc.

Blip streams work very closely with the various switching combinators in Control.Auto.Switch. If anything, if there is only one reason to use Blips, it's with the various switching mechanisms in that module. All of the switching combinators rely on the fact that your Blip streams follow proper semantics, further emphasizing the importance of conforming to the semantics.

For library, framework, and back-end developers

Remember that this module is only meant to export "safe" combinators that try their best to maintain Blip semantics. Think of this module as a useful guideline to help programmers maintain semantics at compile-time, by only exporting not-as-dangerous combinators.

However, all of these rules are for the denotation of your /program logic/. These rules are for the benefit of reasoning about the behavior of your program at the logic level.

As a library or framework or back-end developer, however, you aren't programming at the logic level, but rather at the gritty implementation level. So, you might want to provide Blip streams and for your library users or application developers or the game logic you are writing.

For this, you might find the hidden constructors and tools in Control.Auto.Blip.Internal helpful, and there is more information at the documentation for that module.

data Blip a

A type representing a "discrete" sort of event-like thing. It's supposed to represent something that happens alone, and one at a time, as opposed to things that are "on" or "off" for large intervals at a time.

It's here mainly because it's a pretty useful abstraction in the context of the many combinators found in various modules of this library. If you think of an Auto m a (Blip b) as a "Blip stream", then there are various combinators and functions that are specifically designed to manipulate "Blip streams".

For the purposes of the semantics of what Blip is supposed to represent, its constructors are hidden. (Almost) all of the various Blip combinators (and its very useful Functor instance) "preserve Blipness" --- one-at-a-time occurrences remain one-at-a-time under all of these combinators, and you should have enough so that direct access to the constructor is not needed.

If you are creating a framework, library, or backend, you might want to manually create Blip stream-producing Autos for your users to access. In this case, you can import the constructors and useful internal (and, of course, semantically unsafe) functions from Control.Auto.Blip.Internal.

perBlip :: Monad m => Auto m a b -> Auto m (Blip a) (Blip b)

Takes an Auto m a b (an Auto that turns incoming as into outputting bs) into an Auto m (Blip a) (Blip b); the original Auto is lifted to only be applied to emitted contents of a Blip stream.

When the stream emits, the original Auto is "stepped" with the emitted value; when it does not, it is paused and frozen until the next emission.

>>> let a    = perBlip (sumFrom 0)
>>> let blps = eachAt 2 [1,5,2]
>>> let Output bres _ = stepAutoN' 8 blps ()
>>> bres
[NoBlip, Blip 1, NoBlip, Blip 5, NoBlip, Blip 2, NoBlip, NoBlip]
>>> let Output ares _ = stepAutoN' 8 a ()
>>> ares
[NoBlip, Blip 1, NoBlip, Blip 6, NoBlip, Blip 8, NoBlip, NoBlip]

Merging

merge

Arguments

:: (a -> a -> a)
merging function

-> Blip a

-> Blip a

-> Blip a

Merge two Blips with a merging function. Is only a occuring Blip if *both* Blips are simultaneously occuring.

mergeL :: Blip a -> Blip a -> Blip a infixr 5

Merges two Blip streams together into one, which emits either of the original Blip streams emit. If both emit at the same time, the left (first) one is favored.

mergeR :: Blip a -> Blip a -> Blip a infixl 5

Merges two Blip streams together into one, which emits either of the original Blip streams emit. If both emit at the same time, the right (second) one is favored.

mergeLs :: [Blip a] -> Blip a

Merge all the blip streams together into one, favoring the first emitted value.

mergeRs :: [Blip a] -> Blip a

Merge all the blip streams together into one, favoring the last emitted value.

foldrB :: (a -> a -> a) -> a -> [Blip a] -> Blip a

Merge all of the blip streams together, using the given merging function associating from the right.

foldlB' :: (a -> a -> a) -> a -> [Blip a] -> Blip a

Merge all of the blip streams together, using the given merging function associating from the left.

Blip stream creation (dangerous!)

emitOn

Arguments

:: (a -> Bool)
predicate to emit on

-> Auto m a (Blip a)

An Auto producing a Blip stream that emits the input whenever the input satisfies a given predicate.

Warning! This Auto has the capability of "breaking" Blip semantics. Be sure you know what you are doing when using this. Blip streams are semantically supposed to only emit at discrete, separate occurrences. Do not use this for interval-like (on and off for chunks at a time) things; each input should be dealt with as a separate thing.

For interval semantics, we have Interval from Control.Auto.Interval.

Good example:

-- is only emitting at discrete blips
emitOn even . iterator (+ 1) 0

Bad examples:

-- is emitting for "durations" or "intervals" of time.
emitOn (< 10) . iterator (+ 1) 0

emitOn (const True) . foo

These bad examples would be good use cases of Interval.

emitJusts

Arguments

:: (a -> Maybe b)
"predicate" to emit on.

-> Auto m a (Blip b)

An Auto that runs every input through a a -> Maybe b test and produces a blip stream that emits the value inside every Just result.

A less "boolean-blind" version of emitOn.

Warning! Carries all of the same dangers of emitOn. You can easily break Blip semantics with this if you aren't sure what you are doing. Remember to only emit at discrete, separate occurences, and not for interval-like (on and off for chunks at a time) things. For interval semantics, we have Control.Auto.Interval.

See the examples of emitOn for more concrete good/bad use cases.

onJusts :: Auto m (Maybe a) (Blip a)

An Auto that emits whenever it receives a Just input, with the value inside the Just.

Warning! Carries all of the same dangers of emitOn. You can easily break Blip semantics with this if you aren't sure what you are doing. Remember to only emit at discrete, separate occurences, and not for interval-like (on and off for chunks at a time) things. For interval semantics, we have Control.Auto.Interval.

See the examples of emitOn for more concrete good/bad use cases.

Blip stream collapse

fromBlips

Arguments

:: a
the "default value" to output when the input is not emitting.

-> Auto m (Blip a) a

fromBlips d is an Auto that decomposes the incoming Blip stream by constantly outputting d except when the stream emits, and outputs the emitted value when it does.

holdWith :: Serialize a => a -> Auto m (Blip a) a

holdWith y0 is an Auto that always outputs the /most recently emitted value from the input Blip/ stream. Before anything is emitted, y0 is outputted as a placeholder.

Contrast with hold from Control.Auto.Interval.

holdWith_ :: a -> Auto m (Blip a) a

A non-serializing/non-resumable version of holdWith.

Step/"time" based Blip streams and generators

never :: Auto m a (Blip b)

An Auto that ignores its input and produces a Blip stream that never emits.

immediately :: Auto m a (Blip a)

An Auto that produces a Blip stream that emits on the first ever "tick", with the input at that step. It never emits again after that.

Often used with pure:

>>> immediately . pure "Emit me!"

Or, in proc notation:

>>> blp <- immediately -< "Emit me!"

To get a Blip stream that emits a given value (eg., "Emit me!") once and stops emitting ever again.

inB

Arguments

:: Int
number of steps before value is emitted.

-> Auto m a (Blip a)

An Auto that produces a Blip stream that only emits once, on the given number step. It emits the input on that many steps.

immediately == inB 1

every

Arguments

:: Int
emit every n steps.

-> Auto m a (Blip a)

every n is an Auto that takes in a stream of as and produces a blip stream that emits the input a every n steps. First emission is on the nth step.

Will usually break Blip semantics when you pass in 1.

eachAt

Arguments

:: Serialize b

=> Int
emit every n steps

-> [b]
list to emit values from

-> Auto m a (Blip b)

eachAt n xs is an Auto that ignores its input and emits each element of xs one at a time, evey n steps. First emitted value is at step n.

Once the list is exhausted, never emits again.

Obviously breaks Blip semantics when you pass in 1.

The process of serializing and resuming this Auto is O(n) space and time with the length of xs. So don't serialize this if you plan on passing an infinite list :)

eachAt_

Arguments

:: Int
emit every n steps

-> [b]
list to emit values from

-> Auto m a (Blip b)

The non-serializing/non-resumable version of eachAt.

Modifying Blip streams

tagBlips

Arguments

:: b
value to replace every emitted value with

-> Auto m (Blip a) (Blip b)

Re-emits every emission from the input Blip stream, but replaces its value with the given value.

tagBlips x == modifyBlips (const x)

modifyBlips

Arguments

:: (a -> b)
function to modify emitted values with

-> Auto m (Blip a) (Blip b)

Re-emits every emission from the input Blip stream, but applies the given function to the emitted value.

(<&) :: Monad m => Auto m a (Blip b) -> Auto m a (Blip b) -> Auto m a (Blip b) infixr 5

Takes two Autos generating Blip streams and returns a "merged" Auto that emits when either of the original Autos emit. When both emit at the same time, the left (first) one is favored.

a1 <& a2 == mergeL <$> a1 <*> a2

(&>) :: Monad m => Auto m a (Blip b) -> Auto m a (Blip b) -> Auto m a (Blip b) infixl 5

Takes two Autos generating Blip streams and returns a "merged" Auto that emits when either of the original Autos emit. When both emit at the same time, the right (second) one is favored.

a1 &> a2 == mergeR <$> a1 <*> a2

once :: Auto m (Blip a) (Blip a)

Supress all emissions except for the very first.

notYet :: Auto m (Blip a) (Blip a)

Suppress only the first emission, and ignore the rest.

filterB

Arguments

:: (a -> Bool)
filtering predicate

-> Auto m (Blip a) (Blip a)

Suppress all emissions when the predicate is false.

mapMaybeB :: (a -> Maybe b) -> Auto m (Blip a) (Blip b)

Maps the given function onto every emission, and suppresses all those whose results are Nothing.

takeB

Arguments

:: Int
number of emissions to allow to pass

-> Auto m (Blip a) (Blip a)

takeB n allows only the first n emissions to pass.

takeWhileB

Arguments

:: (a -> Bool)
filtering predicate

-> Auto m (Blip a) (Blip a)

Allow emissions to pass until the first that fails the predicate.

dropB

Arguments

:: Int
number of emissions to suppress initially

-> Auto m (Blip a) (Blip a)

dropB n suppresses the first n emissions and passes through the rest.

dropWhileB

Arguments

:: (a -> Bool)
filtering predicate

-> Auto m (Blip a) (Blip a)

Suppress all emissions until the first one satisfying the predicate, then allow the rest through.

Scanning & Accumulating Blip streams

accumB

Arguments

:: Serialize b

=> (b -> a -> b)
folding function

-> b
initial value

-> Auto m (Blip a) (Blip b)

Accumulates all emissions in the incoming Blip stream with a "folding function", with a given starting value. b -> a -> b, with a starting b, gives Auto m (Blip a) (Blip b).

The resulting Blip stream will emit every time the input stream emits, but with the "accumulated value".

Basically mkAccum, but on blip stream emissions.

accumB f x0 == perBlip (mkAccum f x0)

accumB_

Arguments

:: (b -> a -> b)
folding function

-> b
initial value

-> Auto m (Blip a) (Blip b)

The non-serializing/non-resuming version of accumB.

scanB

Arguments

:: Serialize b

=> (b -> a -> b)
folding function

-> b
initial value

-> Auto m (Blip a) b

At every step, outputs the result of applying the "folding function" with against the contents of every emitted value so far.

scanB f x0 == (hold . accumB f x0) <|!> pure x0

>>> let a = scanB (+) 0 . eachAt 2 [1,2,3]
>>> let Output res _ = stepAutoN' 8 ()
>>> res
[0, 1, 1, 3, 3, 6, 6, 6, 6]

scanB_

Arguments

:: (b -> a -> b)

-> b
folding function

-> Auto m (Blip a) b
initial value

The non-serializing/non-resuming version of scanB.

mscanB :: (Monoid a, Serialize a) => Auto m (Blip a) a

At every step, outputs the monoid sum of all emitted values received so far.

mscanB_ :: Monoid a => Auto m (Blip a) a

The non-serializing/non-resuming version of mscanB.

countB :: Auto m (Blip a) Int

Outputs the number of emitted values of the input Blip stream so far.

Blips on edges

onChange :: (Serialize a, Eq a) => Auto m a (Blip a)

Produces a blip stream that emits whenever the input value changes. Emits with the new value.

Warning: Note that, when composed on a value that is never expected to keep the same value twice, this technically breaks Blip semantics.

onChange_ :: Eq a => Auto m a (Blip a)

The non-serializing/non-resumable version of onChange.

became

Arguments

:: Serialize a

=> (a -> Bool)
change condition

-> Auto m a (Blip a)

Produces a blip stream that emits whenever the predicate applied to the input switches from false to true. Emits with the triggering input value.

became_

Arguments

:: Monad m

=> (a -> Bool)
change condition

-> Auto m a (Blip a)

The non-serializing/non-resumable version of became.

became'

Arguments

:: Monad m

=> (a -> Bool)
change condition

-> Auto m a (Blip ())

Like became, but emits a '()' instead of the triggering input value.

noLonger

Arguments

:: Serialize a

=> (a -> Bool)
change condition

-> Auto m a (Blip a)

Produces a blip stream that emits whenever the predicate applied to the input switches from true to false. Emits with the triggering input value.

noLonger_

Arguments

:: Monad m

=> (a -> Bool)
change condition

-> Auto m a (Blip a)

The non-serializing/non-resumable version of noLonger.

noLonger'

Arguments

:: Monad m

=> (a -> Bool)
change condition

-> Auto m a (Blip ())

Like noLonger, but emits a '()' instead of the triggering input value.

onFlip

Arguments

:: (Serialize a, Monad m)

=> (a -> Bool)
change condition

-> Auto m a (Blip a)

Produces a blip stream that emits whenever the predicate applied to the input switches from true to false or false to true. Emits with the triggering input value.

onFlip_

Arguments

:: Monad m

=> (a -> Bool)
change condition

-> Auto m a (Blip a)

The non-serializing/non-resumable version of onFlip.

onFlip'

Arguments

:: Monad m

=> (a -> Bool)
change condition

-> Auto m a (Blip Bool)

Like onFlip, but emits a '()' instead of the triggering input value.