Control.Auto.Collection

Description

The Autos in this module are all dedicated to managing and working with "collections" of Autos: Autos that can contain and manage and feed/multiplex input through several internal Autos, and collect the results in the end. These are particularly useful for managing collection of Autos that can be added to or deleted from, like monsters in a map, or bullets.

A lot of these Autos take advantaage Interval semantics (Maybe for continuous on/off periods) to signal when they want to be removed or turned off.

For these, the best way to learn them is probably by seeing examples. However, if there is a time when you might want collections of things that can be added to or removed from dynamically, this might be what you are looking for.

Static collections

zipAuto

Arguments

:: Monad m

=> a
default input value

-> [Auto m a b]
Autos to zip up

-> Auto m [a] [b]

Give a list of Auto m a b and get back an Auto m [a] [b] --- take a list of a's and feed them to each of the Autos, and collects their output b's.

If the input list doesn't have enough items to give to all of the Autos wrapped, then use the given default value. Any extra items in the input list are ignored.

For an example, we're going to make a list of Autos that output a running sum of all of their inputs, but each starting at a different beginning value:

    summerList :: [Auto' Int Int]
    summerList = map sumFrom [0, 10, 20, 30]

Then, let's throw it into zipAuto with a sensible default value, 0:

    summings0 :: Auto' [Int] [Int]
    summings0 = zipAuto 0 summerList

Now let's try it out!

>>> let Output r1 summings1 = stepAuto' summings0 [1,2,3,4]
>>> r1
[ 1, 12, 23, 34]
>>> let Output r2 summings2 = stepAuto' summings1 [5,5]
>>> r2
[ 6, 17, 23, 34]
>>> let Output r3 _         = stepAuto' summings2 [10,1,10,1,10000]
>>> r3
[16, 18, 33, 35]

dZipAuto

Arguments

:: (Serialize a, Monad m)

=> a
default input value

-> [Auto m a b]
Autos to zip up

-> Auto m [a] [b]

Like zipAuto, but delay the input by one step, using the default value as the delaying value.

Let's try the example from zipAuto, except with dZipAuto instead:

    summerList :: [Auto' Int Int]
    summerList = map sumFrom [0, 10, 20, 30]

    summings0 :: Auto' [Int] [Int]
    summings0 = dZipAuto 0 summerList

Trying it out:

>>> let Output r1 summings1 = stepAuto' summings0 [1,2,3,4]
>>> r1
[ 0, 10, 20, 30]
>>> let Output r2 summings2 = stepAuto' summings1 [5,5]
>>> r2
[ 1, 12, 23, 34]
>>> let Output r3 summings3 = stepAuto' summings2 [10,1,10,1,10000]
>>> r3
[ 6, 17, 23, 34]
>>> let Output r4 _         = stepAuto' summings3 [100,100,100,100]
>>> r4
[16, 18, 33, 35]

dZipAuto_

Arguments

:: Monad m

=> a
default input value

-> [Auto m a b]
Autos to zip up

-> Auto m [a] [b]

The non-serializing/non-resuming version of dZipAuto.

zipAutoB

Arguments

:: Monad m

=> [Auto m (Blip a) b]
Autos to zip up

-> Auto m [Blip a] [b]

Takes a bunch of Autos that take Blip streams, and turns them into an Auto that takes a bunch of Blip streams and feeds them into each one in order.

It's basically like zipAuto, except instead of taking in normal streams of values, it takes in Blip streams of values.

If the input streams ever number less than the number of Autos zipped, then the other Autos are just stepped with a Blip stream that doesn't emit in that step.

dZipAutoB

Arguments

:: (Serialize a, Monad m)

=> [Auto m (Blip a) b]
Autos to zip up

-> Auto m [Blip a] [b]

A delayed version of zipAutoB

dZipAutoB_

Arguments

:: Monad m

=> [Auto m (Blip a) b]
Autos to zip up

-> Auto m [Blip a] [b]

The non-serializing/non-resuming version of dZipAutoB.

Dynamic collections

dynZip_ :: Monad m => a -> Auto m ([a], Blip [Interval m a b]) [b]

A dynamic box of Autos. Takes a list of inputs to feed to each one, in the order that they were added. Also takes a Blip stream, which emits with new Autos to to the box.

Add new Autos to the box however you want with the Blip stream. It collects all of the Just outputs and outputs them all. Whenever any of the wrapped Autos begins to output Nothing, it is removed from the box.

The adding/removing aside, the routing of the inputs (the first field of the tuple) to the internal Autos and the outputs behaves the same as with zipAuto.

This will be a pretty powerful collection if you ever imagine adding and destroying behaviors dynamically...like spawning new enemies, or something like that.

Let's see an example...here we are going to be throwing a bunch of Autos that count to five and then die into our dynZip_...once every other step.

    -- count upwards, then die when you reach 5
    countThenDie :: Interval' () Int
    countThenDie = onFor 5 . iterator (+1) 1

    -- emit a new countThenDie every two steps
    throwCounters :: Auto' () (Blip [Interval' () Int])
    throwCounters = tagBlips [countThenDie] . every 2

    a :: Auto' () [Int]
    a = proc _ -> do
        newCounter <- throwCounters -< ()
        dynZip_ ()  -< (repeat (), newCounter)

    -- or
    a' = dynZip_ () . (pure (repeat ()) &&& throwCounters)

>>> let (res, _) = stepAutoN' 15 a ()
>>> res
[[], [1            ]
   , [2,           ]
   , [3, 1         ]
   , [4, 2         ]
   , [5, 3, 1      ]
   , [   4, 2      ]
   , [   5, 3, 1   ]
   , [      4, 2   ]
   , [      5, 3, 1]
]

This is a little unweildy, because Autos maybe disappearing out of the thing while you are trying to feed inputs into it. You might be feeding an input to an Auto...but one of the Autos before it on the list has disappeared, so it accidentally goes to the wrong one.

Because of this, it is suggested that you use dynMap_, which allows you to "target" labeled Autos with your inputs.

TODO: Consider putting it in and running inputs immediately?

dynMap_

Arguments

:: Monad m

=> a
"default" input to feed in

-> Auto m (IntMap a, Blip [Interval m a b]) (IntMap b)

A dynamic box of Autos, indexed by an Int. Takes an IntMap of inputs to feed into their corresponding Autos, and collect all of the outputs into an output IntMap.

Whenever any of the internal Autos return Nothing, they are removed from the collection.

Toy examples here are of limited use, but let's try it out. Here we will have a dynMap_ that feeds each internal Auto back to itself. The result of each is sent directly back to itself.

    looper :: Auto' () (IntMap Int)
    looper = proc _ -> do
        initAutos immediately -< sumFromD <$ [1,-1,0]
        rec
            outs <- dynMap_ 0 . delay mempty -< (outs, initAutos)
        id -< outs

>>> let (res, _) = stepAutoN' 20 looper ()
>>> tail res
[ fromList []
, fromList [(0, 1 ), (1, -1 ), (2, 0 )]
, fromList [(0, 1 ), (1, -1 ), (2, 0 )]
, fromList [(0, 2 ), (1, -2 ), (2, 0 )]
, fromList [(0, 3 ), (1, -3 ), (2, 0 )]
, fromList [(0, 5 ), (1, -5 ), (2, 0 )]
, fromList [(0, 8 ), (1, -8 ), (2, 0 )]
, fromList [(0, 13), (1, -13), (2, 0 )]
, fromList [(0, 21), (1, -21), (2, 0 )]
, fromList [(0, 34), (1, -34), (2, 0 )]
]

The first part of each tuple is the Int "key", and the second part is the value.

Basically, the twice-delayed results are fed into the each corresponding Auto --- the results from the Auto starting at positive 1 is fed back to the Auto with a positive counter....results from the Auto starting at negative 1 are fed back into the Auto with the negative counter...and results of the Auto starting at 0 is fed back into the Auto that stays at 0.

TODO: put in pointers to real examples

Multiplexers

Single input, single output

mux

Arguments

:: (Serialize k, Ord k, Monad m)

=> (k -> Auto m a b)
function to create a new Auto if none at that key already exists.

-> Auto m (k, a) b

Auto multiplexer. Takes a function mapping a key k to an initial Auto to be stored at that key. Then, every turn, takes a key-input pair (k, x) :: (k, a), and "runs" the Auto at that key k with that input value x, and outputs the result.

TODO: talk about how you can "restart" arrows using combinators from switch and stuff.

TODO: keys to use with adding things one after another, like in dynMap.

TODO: example

mux_

Arguments

:: (Ord k, Monad m)

=> (k -> Auto m a b)
function to create a new Auto if none at that key already exists

-> Auto m (k, a) b

The non-serializing/non-resuming version of mux.

Multiple input, multiple output

muxMany

Arguments

:: forall m a b k . (Serialize k, Ord k, Monad m)

=> (k -> Auto m a b)
function to create a new Auto if none at that key already exists

-> Auto m (Map k a) (Map k b)

Auto multiplexer, like mux, but with multple inputs and outputs at a time. Instead of giving in a single key-value pair of type (k, a), give an entire key-value collection, Map k a. It runs all of the given a inputs with their corresponding Autos indexed by k, and outputs all of the results in a Map k b.

TODO: Example

muxMany_

Arguments

:: forall m a b k . (Ord k, Monad m)

=> (k -> Auto m a b)
function to create a new Auto if none at that key already exists

-> Auto m (Map k a) (Map k b)

The non-serializing/non-resuming version of muxMany.

Function-based

Single input/output

muxF :: forall m a b k c. (Serialize k, Serialize c, Ord k, Monad m) => (k -> Maybe c -> Auto m a (Maybe b)) -> Auto m (k, Either (c, a) a) (Maybe b)

muxF_ :: forall m a b k c. (Ord k, Monad m) => (k -> Maybe c -> Auto m a (Maybe b)) -> Auto m (k, Either (c, a) a) (Maybe b)

The non-serializing/non-resuming version of muxFI.

Parallel input/output

muxFMany :: forall m a b c k. (Serialize k, Serialize c, Ord k, Monad m) => (k -> Maybe c -> Auto m a (Maybe b)) -> Auto m (Map k (Either (c, a) a)) (Map k b)

muxFMany_ :: forall m a b c k. (Ord k, Monad m) => (k -> Maybe c -> Auto m a (Maybe b)) -> Auto m (Map k (Either (c, a) a)) (Map k b)

The non-serializing/non-resuming version of muxFManyI.

Gathering/accumulating collections

Single input, multiple output

gather

Arguments

:: (Ord k, Monad m, Serialize k, Serialize b)

=> (k -> Interval m a b)
function to create a new Auto if none at that key already exists

-> Auto m (k, a) (Map k b)

Keeps a whole bunch of Autos in a Map, and every step, outputs a Map containing each key and the last (Just) value outputted by the Auto stored at that key.

Gets as input a key-value pair (k, x) :: (k, a), and feeds that input to the Auto stored at that key. If there is no Auto yet stored, then generates a fresh one using the given Auto-making function.

As soon as a stored Auto outputs a Nothing, the Auto (and its key) are removed from the Map, and never outputted again in the total output, its most recent value lost forever. However, you can still "recreate" it by passing in a key-value pair with that key; life starts all over again.

TODO: keys to use with adding things one after another, like in dynMap.

TODO: Example

gather_

Arguments

:: (Ord k, Monad m, Serialize k)

=> (k -> Interval m a b)
function to create a new Auto if none at that key already exists

-> Auto m (k, a) (Map k b)

The non-serializing/non-resuming version of gather:

Does serialize the actual Autos themselves; the Autos are all serialized and re-loaded/resumed when 'gather_ f' is resumed.

Does not serialize the "last outputs", so resumed Autos that have not yet been re-run/accessed to get a fresh output are not represented in the output map at first.

gather__

Arguments

:: (Ord k, Monad m)

=> (k -> Interval m a b)
function to create a new Auto if none at that key already exists

-> Auto m (k, a) (Map k b)

The non-serializing/non-resuming vervsion of gather:

Serializes neither the Autos themselves nor the "last outputs" --- essentially, serializes/resumes nothing.

gatherMany

Arguments

:: forall k a m b . (Ord k, Monad m, Serialize k, Serialize b)

=> (k -> Interval m a b)
function to create a new Auto if none at that key already exists

-> Auto m (Map k a) (Map k b)

Like gather, but allows you to input multiple key-value pairs every step, as a Map. Still outputs a Map with "the last result of every stored Auto".

TODO: Example

gatherMany_

Arguments

:: forall k a m b . (Ord k, Monad m, Serialize k)

=> (k -> Interval m a b)
function to create a new Auto if none at that key already exists

-> Auto m (Map k a) (Map k b)

The non-serializing/non-resuming version of gatherMany:

Does serialize the actual Autos themselves; the Autos are all serialized and re-loaded/resumed when 'gatherMany_ f' is resumed.

Does not serialize the "last outputs", so resumed Autos that have not yet been re-run/accessed to get a fresh output are not represented in the output map at first.

gatherMany__

Arguments

:: forall k a m b . (Ord k, Monad m)

=> (k -> Interval m a b)
function to create a new Auto if none at that key already exists

-> Auto m (Map k a) (Map k b)

The non-serializing/non-resuming vervsion of gatherMany:

Serializes neither the Autos themselves nor the "last outputs" --- essentially, serializes/resumes nothing.

Function-based

gatherF :: forall k a m b c. (Ord k, Monad m, Serialize c, Serialize k, Serialize b) => (k -> Maybe c -> Auto m a (Maybe b)) -> Auto m (k, Either (c, a) a) (Map k b)

gatherF_ :: forall k a m b c. (Ord k, Monad m, Serialize c, Serialize k) => (k -> Maybe c -> Auto m a (Maybe b)) -> Auto m (k, Either (c, a) a) (Map k b)

gatherF__ :: forall k a m b c. (Ord k, Monad m) => (k -> Maybe c -> Auto m a (Maybe b)) -> Auto m (k, Either (c, a) a) (Map k b)

gatherFMany :: forall k m a b c. (Ord k, Monad m, Serialize c, Serialize k, Serialize b) => (k -> Maybe c -> Interval m a b) -> Auto m (Map k (Either (c, a) a)) (Map k b)

gatherFMany_ :: forall k m a b c. (Ord k, Monad m, Serialize c, Serialize k) => (k -> Maybe c -> Interval m a b) -> Auto m (Map k (Either (c, a) a)) (Map k b)

gatherFMany__ :: forall k a m b c. (Ord k, Monad m) => (k -> Maybe c -> Interval m a b) -> Auto m (Map k (Either (c, a) a)) (Map k b)