Control.Auto.Core

Description

This module defines and provides the core types, (smart) constructors, and general utilities used by the Auto library.

A lot of low-level functionality is provided here which is most likely unnecessary for most applications; many are mostly for internal usage or advanced/fine-grained usage; it also isn't really enough to do many things with, either. It's recommended that you import Control.Auto instead, which re-organizes the more useful parts of this module in addition with useful parts of others to provide a nice packaged entry point.

For information on how to actually use these types, see Control.Auto.Tutorial.

Auto

Type

data Auto m a b

The Auto type. Basically represents a function containing its own localized internal state. If you have an Auto a b, you can "step" it with stepAuto and an a, to get a b and a "next Auto". The a is the input, and the b is the output, and the next Auto is the Auto with updated internal state.

The "stepping" process can be monadic:

stepAuto :: Auto m a b -> a -> m (Output m a b)

So you can have access to, say, a shared environment using Reader or something like that.

Auto is mostly useful because of its Functor, Applicative, Category, and Arrow (and Arrow-related) instances. These allow you to modify, combine, chain, and side-chain Autos in expressive way, allowing you to build up complex ones from combinations of simple, primitive ones.

TODO: see tutorial

The Auto also contains information on its own serialization, so you can serialize and re-load the internal state without actually accessing it.

type Auto' = Auto Identity

Special case of Auto where the underlying Monad is Identity.

autoConstr :: Auto m a b -> String

Returns a string representation of the internal constructor of the Auto. Useful for debugging the result of compositions and functions and seeing how they affect the internal structure of the Auto.

toArb :: Monad m => Auto m a b -> Auto m a b

Re-structure Auto internals to use the Arb ("arbitrary") constructors, as recursion-based mealy machines. Almost always a bad idea in every conceivable situation. Why is it even here?

Running

stepAuto

Arguments

:: Monad m

=> Auto m a b
the Auto to step

-> a
the input

-> m (Output m a b)
the output, and the updated Auto'.

Runs the Auto through one step.

Remember that at every step for an Auto m a b, you provide an a input and receive a b output with an "updated"/"next" Auto.

>>> let a = mkAccum (+) 0 :: Auto Identity Int Int
            -- an Auto that sums all of its input.
>>> let Identity (Output y a') = stepAuto a 3
>>> y      -- the result
3 :: Int
>>> :t a'   -- the updated 'Auto'
a' :: Auto Identity Int Int

(Identity, from Data.Functor.Identity, is the "dumb Functor": data Identity a = Identity a)

If you think of an Auto m a b as a "stateful function" of type a -> m b, then stepAuto lets you "run" it.

stepAuto'

Arguments

:: Auto' a b
the Auto' to step

-> a
the input

-> Output' a b
the output, and the updated Auto'

stepAuto, but for an Auto' --- the underlying Monad is Identity. Returns the output stripped of Identity.

If you think of an Auto' a b as a "stateful function" of type a -> b, then stepAuto' lets you "run" it.

Serializing

The Auto type offers an interface in which you can serialize ("freeze") and "resume" an Auto, in ByteString (binary) form.

You can "freeze" any Auto into a ByteString using encodeAuto (or, if you want the raw Put (from Data.Serialize) for some reason, there's saveAuto.

You can "resume" any Auto from a ByteString using decodeAuto (or, if you want the raw Get for some reason, there's loadAuto).

Note decodeAuto and loadAuto "resume" a given Auto. That is, if you call decodeAuto on a "fresh Auto", it'll decode a ByteString into that Auto, but "resumed". That is, it'll "fast forward" that Auto into the state it was when it was saved.

For example, let's say I have a = mkAccum (+) 0, the Auto that returns the sum of everything it has received so far. If I feed it 3 and 10, it'll have its internal accumulator as 13, keeping track of all the numbers it has seen so far.

>>> let a             = mkAccum (+) 0
>>> let Output _ a'   = stepAuto' a  3
>>> let Output _ a''  = stepAuto' a' 10

I can then use encodeAuto to "freeze"/"save" the Auto into the ByteString bs:

>>> let bs            = encodeAuto a''

To "resume" / "load" it, I can use decodeAuto to "resume" the original a. Remember, a was our original Auto, the summer Auto with a starting accumulator of 0. We use decodeAuto to "resume" it, with and resume it with its internal accumulator at 13.

>>> let Right resumed = decodeAuto a bs
>>> let Output y _    = stepAuto' resumed 0
13

Note that all of these would have had the same result:

>>> let Right resumed = decodeAuto a'  bs
>>> let Right resumed = decodeAuto a'' bs
>>> let Right resumed = decodeAuto (mkAccum (+) 0) bs

I mean, after all, if decodeAuto "fast forwards" an Auto to the state it was at when it was frozen...then all of these should really be resumed to the same point, right?

One way you can think about it is that loadAuto / decodeAuto takes an Auto and creates a "blueprint" from that Auto, on how to "load it"; the blueprint contains what the form of the internal state is, and their offets in the ByteString. So in the above, a, a', a'', and 'mkAccum (+) 0' all have the same "blueprint" --- their internal states are of the same structure.

Some specific Autos (indicated by a naming convention) might choose to have internal state, yet ignore it when saving/loading. So, saving it actaully saves no state, and "resuming" it really doesn't do anything. That is, decodeAuto a_ bs = Right a_. There isn't a real way to identify from the type of the Auto if it will properly save/resume or not, so you have to keep track of this yourself. In all of the Auto "included" in this library, any Auto whose name does not end in _ will serialize and resume. An Auto whose name ends in _ is taken by naming convention to be a non-resuming Auto.

In your own compositions, if you are sure to always use resuming Autos, your composition will also be properly resuming...so you don't have to worry about this! You shouldn't really ever be "surprised", because you'll always explicitly chose the resuming version for Autos you want to resume, and the non-resuming version for those you don't.

TODO: replace "decode" with "resume".

encodeAuto :: Auto m a b -> ByteString

Encode an Auto and its internal state into a ByteString.

decodeAuto :: Auto m a b -> ByteString -> Either String (Auto m a b)

Resume an Auto from its ByteString serialization, giving a Left if the deserialization is not possible.

saveAuto :: Auto m a b -> Put

Returns a Put --- instructions (from Data.Serialize) on how to "freeze" the Auto, with its internal state, and save it to a binary encoding. It can later be reloaded and "resumed" by 'loadAuto'/'decodeAuto'.

loadAuto :: Auto m a b -> Get (Auto m a b)

Returns a Get from an Auto --- instructions (from Data.Serialize) on taking a ByteString and "restoring" the originally saved Auto, in the originally saved state.

Auto output

data Output m a b

The output of a stepAuto. Contains the "result" value of the stepping (outRes), and the "next Auto", outAuto.

An Auto m a b will produce an Output m a b; when "stepped" with an a, the "result" (outRes) is a b.

Really, you can just think of this as a fancy tuple.

Constructors

Output

Fields

outRes :: b

Result value of a step

outAuto :: Auto m a b

The next Auto

type Output' = Output Identity

Special case of Output where the underlying Monad of outAuto is Identity.

onOutput

Arguments

:: (b -> b')
function over the result

-> (Auto m a b -> Auto m a' b')
function over the resulting Auto

-> Output m a b

-> Output m a' b'

Map two functions onto both fields of an Output.

If you want to map an a -> b onto both fields (the result and the result of the next Auto), you can use the Functor instance instead.

onOutRes

Arguments

:: (b -> b)
function over the result

-> Output m a b

-> Output m a b

Map a function onto the outRes of an Output: the "result" of a stepAuto.

Useful for completely pointless and probably obfuscating point free code :D

onOutAuto

Arguments

:: (Auto m a b -> Auto m a' b)
function over the resulting Auto

-> Output m a b

-> Output m a' b

Map a function onto the 'outAuto of an Output, the resulting Auto. See note on onOutRes.

Auto constructors

Lifting values and functions

mkConst

Arguments

:: b
constant value to be outputted

-> Auto m a b

Construct the Auto that always yields the given value, ignoring its input.

Provided for API constency, but you should really be using pure from the Applicative instance, from Control.Applicative, which does the same thing.

mkConstM

Arguments

:: m b
monadic action to be executed at every step

-> Auto m a b

Construct the Auto that always "executes" the given monadic value at every step, yielding the result and ignoring its input.

Provided for API consistency, but you shold really be using effect from Control.Auto.Effects, which does the same thing.

mkFunc

Arguments

:: (a -> b)
pure function

-> Auto m a b

Construct a stateless Auto that simply applies the given (pure) function to every input, yielding the output.

This is rarely needed; you should be using arr from the Arrow instance, from Control.Arrow.

mkFuncM

Arguments

:: (a -> m b)
"monadic" function

-> Auto m a b

Construct a statelss Auto that simply applies and executes the givne (monadic) function to every input, yielding the output.

It's recommended that you use arrM from Control.Auto.Effects. This is only really provided for consistency.

from State transformers

mkState

Arguments

:: Serialize s

=> (a -> s -> (b, s))
state transformer

-> s
intial state

-> Auto m a b

Construct an Auto from a state transformer: an a -> s -> (b, s) gives you an Auto m a b, for any Monad m. At every step, it takes in the a input, runs the function with the stored internal state, returns the b result, and now contains the new resulting state. You have to intialize it with an initial state, of course.

Try not to use this if it's ever avoidable, unless you're a framework developer or something. Try make something by combining/composing the various Auto combinators.

This version is a wrapper around mkAuto, that keeps track of the serialization and re-loading of the internal state for you, so you don't have to deal with it explicitly.

If your state s does not have a Serialize instance, then you should either write a meaningful one, provide the serialization methods manually with mkState', or throw away serializability and use mkState_.

mkState_

Arguments

:: (a -> s -> (b, s))
state transformer

-> s
initial state

-> Auto m a b

A version of mkState, where the internal state isn't serialized. It can be "saved" and "loaded", but the state is lost in the process.

Useful if your state s cannot have a meaningful Serialize instance.

mkStateM

Arguments

:: Serialize s

=> (a -> s -> m (b, s))
(monadic) state transformer

-> s
initial state

-> Auto m a b

Construct an Auto from a "monadic" state transformer: a -> s -> m (b, s) gives you an Auto m a b. At every step, it takes in the a input, runs the function with the stored internal state and "executes" the m (b, s) to get the b output, and stores the s as the new, updated state. Must be initialized with an initial state.

Try not to use this if it's ever avoidable, unless you're a framework developer or something. Try make something by combining/composing the various Auto combinators.

This version is a wrapper around mkAuto, that keeps track of the serialization and re-loading of the internal state for you, so you don't have to deal with it explicitly.

If your state s does not have a Serialize instance, then you should either write a meaningful one, provide the serialization methods manually with mkStateM', or throw away serializability and use mkStateM_.

mkStateM_

Arguments

:: (a -> s -> m (b, s))
(monadic) state transformer

-> s
initial state

-> Auto m a b

A version of mkStateM, where the internal state isn't serialized. It can be "saved" and "loaded", but the state is lost in the process.

Useful if your state s cannot have a meaningful Serialize instance.

mkState'

Arguments

:: Get s
Get; strategy for reading and deserializing the state

-> (s -> Put)
Put; strategy for serializing given state

-> (a -> s -> (b, s))
state transformer

-> s
intial state

-> Auto m a b

A version of mkState, where the internal state doesn't have a Serialize instance, so you provide your own instructions for getting and putting the state.

mkStateM'

Arguments

:: Get s
Get; strategy for reading and deserializing the state

-> (s -> Put)
Put; strategy for serializing given state

-> (a -> s -> m (b, s))
(monadic) state transformer

-> s
initial state

-> Auto m a b

A version of mkStateM, where the internal state doesn't have a Serialize instance, so you provide your own instructions for getting and putting the state.

from Accumulators

Result-first

mkAccum

Arguments

:: Serialize b

=> (b -> a -> b)
accumulating function

-> b
initial accumulator

-> Auto m a b

Construct an Auto from a "folding" function: b -> a -> b yields an Auto m a b. Basically acts like a foldl or a scanl. There is an internal accumulator that is "updated" with an a at every step. Must be given an initial accumulator.

Example: an Auto that sums up all of its input.

>>> let summer = mkAccum (+) 0
>>> let Output sum1 summer' = stepAuto summer 3
>>> sum1
3
>>> let Output sum2 _       = stepAuto summer'' 10
>>> sum2
13

If your accumulator b does not have a Serialize instance, then you should either write a meaningful one, or throw away serializability and use mkAccum_.

mkAccum_

Arguments

:: (b -> a -> b)
accumulating function

-> b
intial accumulator

-> Auto m a b

A version of 'mkAccum_, where the internal accumulator isn't serialized. It can be "saved" and "loaded", but the state is lost in the process.

Useful if your accumulator b cannot have a meaningful Serialize instance.

mkAccumM

Arguments

:: (Serialize b, Monad m)

=> (b -> a -> m b)
(monadic) accumulating function

-> b
initial accumulator

-> Auto m a b

Construct an Auto from a "monadic" "folding" function: b -> a -> m b yields an Auto m a b. Basically acts like a foldM or scanM (if it existed). here is an internal accumulator that is "updated" with an input a with the result of the executed m b at every step. Must be given an initial accumulator.

If your accumulator b does not have a Serialize instance, then you should either write a meaningful one, or throw away serializability and use mkAccumM_.

mkAccumM_

Arguments

:: Monad m

=> (b -> a -> m b)
(monadic) accumulating function

-> b
initial accumulator

-> Auto m a b

A version of 'mkAccumM_, where the internal accumulator isn't serialized. It can be "saved" and "loaded", but the state is lost in the process.

Useful if your accumulator b cannot have a meaningful Serialize instance.

Initial accumulator-first

mkAccumD

Arguments

:: Serialize b

=> (b -> a -> b)
accumulating function

-> b
initial accumulator

-> Auto m a b

A "delayed" version of mkAccum, where the first output is actually the initial state of the accumulator. Useful in recursive bindings.

mkAccumD_

Arguments

:: (b -> a -> b)
accumulating function

-> b
intial accumulator

-> Auto m a b

The non-resuming/non-serializing version of mkAccumD.

mkAccumMD

Arguments

:: (Serialize b, Monad m)

=> (b -> a -> m b)
(monadic) accumulating function

-> b
initial accumulator

-> Auto m a b

A "delayed" version of mkAccumM, where the first output is actually the initial state of the accumulator. Useful in recursive bindings.

mkAccumMD_

Arguments

:: Monad m

=> (b -> a -> m b)
(monadic) accumulating function

-> b
initial accumulator

-> Auto m a b

The non-resuming/non-serializing version of mkAccumMD.

Arbitrary Autos

mkAuto

Arguments

:: Get (Auto m a b)
resuming/loading Get

-> Put
saving Put

-> (a -> Output m a b)
step function

-> Auto m a b

Construct an Auto by explicity giving its serialization, deserialization, and the (pure) function from a to b and the "next Auto".

Ideally, you wouldn't have to use this unless you are making your own framework. Try your best to make what you want by assembling primtives together.

mkAuto_

Arguments

:: (a -> Output m a b)
step function

-> Auto m a b

Like mkAuto, but without any way of meaningful serializing or deserializing.

Be careful! This Auto can still carry arbitrary internal state, but it cannot be meaningfully serialized or re-loaded/resumed. You can still pretend to do so using 'loadAuto'/'saveAuto'/'encodeAuto'/'decodeAuto' (and the type system won't stop you), but when you try to "resume"/decode it, its state will be lost.

mkAutoM

Arguments

:: Get (Auto m a b)
resuming/loading Get

-> Put
saving Put

-> (a -> m (Output m a b))
(monadic) step function

-> Auto m a b

Construct an Auto by explicitly giving its serializiation, deserialization, and the (monadic) function from a to b and the "next Auto".

Ideally, you wouldn't have to use this unless you are making your own framework. Try your best to make what you want by assembling primtives together.

mkAutoM_

Arguments

:: (a -> m (Output m a b))
(monadic) step function

-> Auto m a b

Like mkAutoM, but without any way of meaningful serializing or deserializing.

Be careful! This Auto can still carry arbitrary internal state, but it cannot be meaningfully serialized or re-loaded/resumed. You can still pretend to do so using 'loadAuto'/'saveAuto'/'encodeAuto'/'decodeAuto' (and the type system won't stop you), but when you try to "resume"/decode it, its state will be reset.

Strictness

forceSerial :: Auto m a b -> Auto m a b

Force the serializing components of an Auto.

forcer :: NFData a => Auto m a a

A special Auto that acts like the id Auto, but forces results as they come through to be fully evaluated, when composed with other Autos.

seqer :: Auto m a a

A special Auto that acts like the id Auto, but forces results as they come through to be evaluated to Weak Head Normal Form, with seq, when composed with other Autos.