Control.Auto.Generate

Description

This module contains various Autos that act as "producers", ignoring their input and producing output through some source or a pure/monadic function.

Constant producers

Are you looking for "constant producers"? Autos that constantly output the same thing, or repeatedly execute the same monadic action and return the result? To keep things clean, they aren't re-exported here. But you'll find the "constant Auto" as pure from the Applicative interface:

pure :: Monad m => b -> Auto m a b

And also the "repeatedly execute and return" Auto as effect from Control.Auto.Effects:

effect :: Monad m => m b -> Auto m a b

From lists

fromList

Arguments

:: Serialize b

=> [b]
list to output element-by-element

-> Interval m a b

Construct an Auto that ignores its input and just continually emits elements from the given list. Ouputs Nothing forever after reaching the end of the list.

Outputs an Intervaled output, which is "on" (Just) as long as the list still has elements to output, and is "off" (Nothing) after it runs out.

Serializes itself by storing the entire rest of the list in binary, so if your list is long, it might take up a lot of space upon storage. If your list is infinite, it makes an infinite binary, so be careful!

Storing: O(n) time and space on length of remaining list

Loading: O(1) time in the number of times the Auto has been stepped + O(n) time in the length of the remaining list.

fromList_

Arguments

:: [b]
list to output element-by-element

-> Interval m a b

The non-resuming/non-serializing version of fromList.

fromLongList

Arguments

:: [b]
list to output element-by-element

-> Interval m a b

A version of fromList that is safe for long or infinite lists, or lists with unserializable elements.

There is a small cost in the time of loading/resuming, which is O(n) on the number of times the Auto had been stepped at the time of saving. This is because it has to drop the n first elements in the list, to "resume" to the proper position.

Storing: O(1) time and space on the length of the remaining list

Loading: O(n) time on the number of times the Auto has been stepped, maxing out at O(n) on the length of the entire input list.

From functions

Iterating

iterator

Arguments

:: Serialize b

=> (b -> b)
iterating function

-> b
starting value and initial output

-> Auto m a b

Analogous to iterate from Prelude. Keeps accumulator value and continually applies the function to the value at every step, outputting the result.

The first result is the initial accumulator value.

>>> let (y, _) = stepAutoN' 10 (iterator (*2) 1) ()
>>> y
[1, 2, 4, 8, 16, 32, 64, 128, 256, 512]

iterator_

Arguments

:: (b -> b)
iterating function

-> b
starting value and initial output

-> Auto m a b

The non-resuming/non-serializing version of iterator.

iteratorM

Arguments

:: (Serialize b, Monad m)

=> (b -> m b)
(monadic) iterating function

-> b
starting value and initial output

-> Auto m a b

Like iterator, but with a monadic function.

iteratorM_

Arguments

:: Monad m

=> (b -> m b)
(monadic) iterating function

-> b
starting value and initial output

-> Auto m a b

The non-resuming/non-serializing version of iteratorM.

Enumerating results of a function

discreteF

Arguments

:: (Enum c, Serialize c)

=> (c -> b)
discrete function

-> c
initial input

-> Auto m a b

Given a function from discrete enumerable values, iterates through all of the results of that function.

Still thinking of a good name for this...

>>> take 10 . streamAuto' (discreteF (^2) 0) $ repeat ()
[0, 1, 4, 9, 16, 25, 36, 49, 64, 81]

discreteF_

Arguments

:: Enum c

=> (c -> b)
discrete function

-> c
initial input

-> Auto m a b

The non-resuming/non-serializing version of discreteF.

Unfolding

"Iterating with state".

unfold

Arguments

:: Serialize c

=> (c -> Maybe (b, c))
unfolding function

-> c
initial accumulator

-> Interval m a b

Analogous to unfoldr from Prelude. Creates a generating Interval by maintaining an internal accumulator of type c and, at every stpe, applying to the unfolding function to the accumulator. If the result is Nothing, then the Interval will turn "off" forever (output Nothing forever); if the result is Just (y, acc), then it will output y and store acc as the new accumulator.

Given an initial accumulator.

unfold_

Arguments

:: (c -> Maybe (b, c))
unfolding function

-> c
initial accumulator

-> Interval m a b

The non-resuming & non-serializing version of unfold.

unfoldM

Arguments

:: (Serialize c, Monad m)

=> (c -> m (Maybe (b, c)))
unfolding function

-> c
initial accumulator

-> Interval m a b

Like unfold, but the unfolding function is monadic.

unfoldM_

Arguments

:: Monad m

=> (c -> m (Maybe (b, c)))
unfolding function

-> c
initial accumulator

-> Interval m a b

The non-resuming & non-serializing version of unfoldM.

Enumerating

enumFromA

Arguments

:: (Serialize b, Enum b)

=> b
initial value

-> Auto m a b

Continually enumerate from the starting value, using succ.

enumFromA_

Arguments

:: Enum b

=> b
initial value

-> Auto m a b

The non-serializing/non-resuming version of enumFromA.