Control.Auto.Interval

Description

This module provides combinators and utilities for working with the semantic concept of "intervals": Autos producing values that can either be "on" or "off"...typically for contiguous chunks at a time.

Intervals

This concept of "on or offness" is represented as an "interval-producing Auto" using Maybe:

    Auto m a (Maybe b)

(In contrast to the "normal", "always-on" Auto, Auto m a b)

Which from now on, we will be referring to with the equivalent /type synonym/:

    type Interval m a b = Auto m a (Maybe b)

So, conceptually, an Interval m a b is the same (from the compiler's point of view) as an Auto m a (Maybe b). If you see it, you can substitute it in your head if it makes it easier.

Normally, an Auto m a b takes in a stream of as and produces a stream of bs. So an Interval m a b takes in a stream of as and produces an intervaled stream of b's --- a stream that is on or off for contiguous chunks at a time.

When it outputs Nothing, it's interpreted as "off"; when it outputs Just x, it's interpreted as "on" with a value of x.

For example, take onFor :: Int -> Interval m a a. onFor n is "on", and lets all values pass through through, for n steps; then it turns "off" forever.

>>> let a        = onFor 2 . count
>>> let (res, _) = stepAutoN' 5 a ()
>>> res
[Just 1, Just 2, Nothing, Nothing, Nothing]

(Recall that count, from Control.Auto.Time, outputs the current step count)

onFor 2 is "on" for two steps, and lets the output of count through; then it is "off", and "blocks" all of the output of count.

Motivation

Intervals happen to particularly useful when used with the various switching combinators from Control.Auto.Switch.

You might find it useful to "sequence" Autos such that they "switch" from one to the other, dynamically. For example, an Auto that acts like pure 0 for three steps, and then like count for the rest:

>>> let a1 = (onFor 3 . pure 0) --> count
>>> let (res1, _) = stepAutoN' 8 a1 ()
>>> res1
[0, 0, 0, 1, 2, 3, 4, 5]

(Recall that pure x is the Auto that ignores its input and always outputs x)

Or in reverse, an Auto that behaves like count until the count is above 3, then switches to pure 0

>>> let a2 = (when (<= 3) . count) --> pure 0
>>> let (res2, _) = stepAutoN' 8 a2 ()
>>> res2
[1, 2, 3, 0, 0, 0, 0, 0]

That's just a small example using one switching combinator, -->. But hopefully it demonstrates that one powerful motivation behind "intervals" being a "thing" is because of how it works with switches.

Another neat motivation is that intervals work pretty well with the Blip semantic tool, as well.

The following Interval will be "off" and suppress all of its input (from count) until the Blip stream produced by inB 3 emits something, then it'll allow count to pass.

>>> let a3        = after . (count &&& inB 3)
>>> let (res3, _) = stepAutoN' 5 a3 ()
>>> res3
[Nothing, Nothing, Just 3, Just 4, Just 4]

The Contract

So, why have an Interval type, and not always just use Auto?

You can say that, if you are given an Interval, then it comes with a "contract" (by documentation) that the Auto will obey /interval semantics/.

Auto m a (Maybe b) can mean a lot of things and represent a lot of things.

However, if you offer something of an Interval type, or if you find something of an Interval type, it comes with some sort of assurance that that Auto will behave like an interval: on and off for contiguous periods of time.

In addition, this allows us to further clarify /what our functions expect/. By saying that a function expects an Interval:

    chooseInterval :: [Interval m a b]
                   -> Interval m a b

chooseInterval has the ability to "state" that it expects things that follow interval semantics in order to "function" properly and in order to properly "return" an Interval.

Of course, this is not enforced by the compiler. However, it's useful to create a way to clearly state that what you are offering or what you are expecting does indeed follow this useful pattern.

Combinators

Converting back into normal streams

You can convert interval streams back into normal streams by using fromInterval and fromIntervalWith, analogous to fromMaybe and maybe from Data.Maybe, respectively:

>>> let a        = fromIntervalWith "off" show . onFor 2 . count
>>> let (res, _) = stepAutoN' 5 a ()
>>> res
["1", "2", "off", "off", "off"]

Choice

You can also "choose" between interval streams, with choice combinators like <|?> and <|!>.

>>> let a = onFor 2 . pure "hello"
       <|!> onFor 4 . pure "world"
       <|!> pure "goodbye!"
>>> let (res, _) = stepAutoN' 6 a ()
>>> res
["hello", "hello", "world", "world", "goodbye!", "goodbye!"]

The above could also be written with choose:

>>> let a = choose (pure "goodbye!")
                   [ onFor 2 . pure "hello"
                   , onFor 4 . pure "world"
                   ]

Composition

Another tool that makes Intervals powerful is the ability to compose them.

If you have an Auto m a b and an Auto m b c, then you can compose them with ..

If you have an Auto m a b and an Interval m b c, then you can compose them by throwing in a toOn in the chain, or fmap Just:

    a               :: Auto m a b
    i               :: Interval m b c
    i . toOn . a    :: Interval m a c
    fmap Just a :: Interval m a b
    i . fmap Just a :: Interval m a c

If you have an Interval m a b and an Auto m b c, you can "lift" the second Auto to be an Auto that only "acts" on "on"/Just outputs of the Interval:

    i            :: Interval m a b
    a            :: Auto m b c
    during a     :: Auto m (Maybe a) (Maybe b)
    during a . i :: Interval m a c

Finally, the kleisli composition: if you have an Interval m a b and an Interval m b c, you can use compI: (or also bindI)

    i1            :: Interval m a b
    i2            :: Interval m b c
    i2 `'compI'` i1 :: Interval m a b c

>>> let a1        = onFor 4 `compI` offFor 1 . count
>>> let (res1, _) = stepAutoN' 6 a1 ()
>>> res1
[Nothing, Just 2, Just 3, Just 4, Nothing, Nothing]

>>> let a2 = when even `compI` onFor 4 `compI` offFor 1 . count
>>> let (res2, _) = stepAutoN' 6 a2 ()
>>> res2
[Nothing, Just 2, Nothing, Just 4, Nothing, Nothing]

The implementation works so that any "on"/Just inputs will step the lifted Auto like normal, with the contents of the Just, and any "off"/Nothing inputs cause the lifted Auto to be skipped and frozen.

compI adds a lot of power to Interval because now you can always work "with Intervals", bind them just like normal Autos, and then finally "exit" them after composing and combining many.

Warning: Switching

Note that when any of these combinators "block" (or "inhibit" or "suppress", whatever you call it) their input as a part of a composition pipeline (as in for off, onFor, offFor, etc.), the input Autos are still stepped and "run". If the inputs had any monad effects, they would too be exected at every step. In order to "freeze" and not run or step an Auto at all, you have to use switches.

(during and bindI are not included in this bunch.)

type Interval m a b = Auto m a (Maybe b)

An Interval is "just" a type alias for Auto m a (Maybe b). If you ended up here with a link...no worries! If you see Interval m a b, just think Auto m a (Maybe b)!

type Interval' a b = Auto' a (Maybe b)

Static Intervals

off :: Interval m a b

An Auto that produces an interval that always "off" (Nothing), never letting anything pass.

Note that any monadic effects of the input Auto when composed with off are still executed, even though their result value is suppressed.

off == arr (const Nothing)

toOn :: Interval m a a

An Auto that takes a value stream and turns it into an "always-on" interval, with that value. Lets every value pass through.

toOn == arr Just

fromInterval

Arguments

:: a
value to output for "off" periods

-> Auto m (Maybe a) a

An Auto taking in an interval stream and transforming it into a normal value stream, using the given default value whenever the interval is off/blocking.

fromInterval d = arr (fromMaybe d)

fromIntervalWith :: b -> (a -> b) -> Auto m (Maybe a) b

An Auto taking in an interval stream and transforming it into a normal value stream, using the given default value whenever the interval is off/blocking, and applying the given function to the input when the interval is on/passing. Analogous to maybe from Prelude and Data.Maybe.

fromIntervalWith d f = arr (maybe d f)

onFor

Arguments

:: Int
amount of steps to stay "on" for

-> Interval m a a

An Auto that behaves like toOn (letting values pass, "on") for the given number of steps, then otherwise is off (suppressing all input values from passing) forevermore.

offFor

Arguments

:: Int
amount of steps to be "off" for.

-> Interval m a a

An Auto that is off for the given number of steps, suppressing all input values, then behaves like toOn forevermore, passing through values as "on" values.

Filter Intervals

when

Arguments

:: (a -> Bool)
interval predicate

-> Interval m a a

An Auto that allows values to pass whenever the input satisfies the predicate...and is off otherwise.

>>> let a        = when (\x -> x >= 2 && x <= 4) . count
>>> let (res, _) = stepAutoN' 6 a ()
>>> res
[Nothing, Just 2, Just 3, Just 4, Nothing, Nothing]

(count is the Auto that ignores its input and outputs the current step count at every step)

unless

Arguments

:: (a -> Bool)
interval predicate

-> Interval m a a

Like when, but only allows values to pass whenever the input does not satisfy the predicate. Blocks whenever the predicate is true.

>>> let a        = unless (\x -> x < 2 &&& x > 4) . count
>>> let (res, _) = stepAutoN' 6 a ()
>>> res
[Nothing, Just 2, Just 3, Just 4, Nothing, Nothing]

(count is the Auto that ignores its input and outputs the current step count at every step)

Choice

(<|!>) infixr 3

Arguments

:: Monad m

=> Interval m a b
interval Auto

-> Auto m a b
"normal" Auto

-> Auto m a b

Chooses between an interval-producing Auto and an "normal" value, "always on" Auto. Behaves like the "on" value of the first Auto if it is on; otherwise, behaves like the second.

>>> let a1        = (onFor 2 . pure "hello") <|!> pure "world"
>>> let (res1, _) = stepAutoN' 5 a1 ()
>>> res1
["hello", "hello", "world", "world", "world"]

This one is neat because it associates from the right, so it can be "chained":

>>> let a2 = onFor 2 . pure "hello"
        <|!> onFor 4 . pure "world"
        <|!> pure "goodbye!"
>>> let (res2, _) = stepAutoN' 6 a2 ()
>>> res2
["hello", "hello", "world", "world", "goodbye!", "goodbye!"]

 a <|!> b <!|> c

associates as

 a <|!> (b <|!> c)

So using this, you can "chain" a bunch of choices between intervals, and then at the right-most, "final" one, provide the default behavior.

Warning: If your underlying monad produces effects, remember that both Autos are run at every step, along with any monadic effects, regardless of whether they are "on" or "off".

(<|?>) infixr 3

Arguments

:: Monad m

=> Interval m a b
choice 1

-> Interval m a b
choice 2

-> Interval m a b

This "chooses" between two interval-producing Autos; behaves like the first Auto if it is "on"; otherwise, behaves like the second.

>>> let a        = (onFor 2 . pure "hello") <|?> (onFor 4 . pure "world")
>>> let (res, _) = stepAutoN' 5 a ()
>>> res
[Just "hello", Just "hello", Just "world", Just "world", Nothing]

You can drop the parentheses, because of precedence; the above could have been written as:

>>> let a' = onFor 2 . pure "hello" <|?> onFor 4 . pure "world"

Warning: If your underlying monad produces effects, remember that both Autos are run at every step, along with any monadic effects, regardless of whether they are "on" or "off".

Note that more often than not, <|!> is probably more useful. This is useful only in the case that you really, really want an interval at the end of it all.

chooseInterval

Arguments

:: Monad m

=> [Interval m a b]
the Autos to run and choose from

-> Interval m a b

Run all Autos from the same input, and return the behavior of the first one that is not Nothing. If all are Nothing, output Nothing.

chooseInterval == foldr (<|?>) off

choose

Arguments

:: Monad m

=> Auto m a b
the Auto to behave like if all others are Nothing

-> [Interval m a b]
Autos to run and choose from

-> Auto m a b

Run all Autos from the same input, and return the behavior of the first one that is not Nothing; if all are Nothing, return the behavior of the "default case".

choose == foldr (<|!>)

Blip-based Intervals

after :: Interval m (a, Blip b) a

Takes in a value stream and a Blip stream. Doesn't allow any values in at first, until the Blip stream emits. Then, allows all values through as "on" forevermore.

>>> let a        = after . (count &&& inB 3)
>>> let (res, _) = stepAutoN' 5 a ()
>>> res
[Nothing, Nothing, Just 3, Just 4, Just 4]

(count is the Auto that ignores its input and outputs the current step count at every step, and inB 3 is the Auto generating a Blip stream that emits at the third step.)

Be careful to remember that after does not actually "switch" anything. In the above example, count is still "run" at every step, and is progressed (and if it were an Auto with monadic effects, they would still be executed). It just isn't allowed to pass through after until the Blip stream emits.

before :: Interval m (a, Blip b) a

Takes in a value stream and a Blip stream. Allows all values through, as "on", until the Blip stream emits...then doesn't let anything pass after that.

>>> let a        = before . (count &&& inB 3)
>>> let (res, _) = stepAutoN' 5 a ()
>>> res
[Just 1, Just 2, Nothing, Nothing, Nothing]

(count is the Auto that ignores its input and outputs the current step count at every step, and inB 3 is the Auto generating a Blip stream that emits at the third step.)

Be careful to remember that before doesn't actually "switch" anything. In the above example, count is still "run" at every step (and if it were an Auto with monad effects, they would still be executed). It's just that the values are suppressed.

between :: Interval m (a, (Blip b, Blip c)) a

Takes in a value stream and two Blip streams. Starts off as "off", not letting anything pass. When the first Blip stream emits, it toggles onto the "on" state and lets everything pass; when the second Blip stream emits, it toggles back onto the "off" state.

>>> let a        = before . (count &&& (inB 3 &&& inB 5))
>>> let (res, _) = stepAutoN' 7 a ()
>>> res
[Nothing, Nothing, Just 3, Just 4, Nothing, Nothing, Nothing]

hold :: Serialize a => Interval m (Blip a) a

Takes in a Blip stream and constantly outputs the last emitted value. Starts off as Nothing.

>>> let a1        = hold . inB 3 . count
>>> let (res1, _) = stepAutoN' 5 a1 ()
>>> res1
[Nothing, Nothing, Just 3, Just 3, Just 3]

You can make this behave as an Auto m (Blip a) b (no possible Nothings) by providing a "default" value, to be used when the input stream has not yet emitted, in one of two ways:

The first, using <|!>:

>>> let a2        = (hold . inB 3 . count) <|!> pure 100
>>> let (res2, _) = stepAutoN' 5 a2 ()
>>> res2
[100, 100, 3, 3, 3]

The second, using fromInterval:

>>> let a3        = fromInterval 100 . hold . inB 3 . count
>>> let (res3, _) = stepAutoN' 5 a3 ()
>>> res3
[100, 100, 3, 3, 3]

hold_ :: Interval m (Blip a) a

The non-serializing/non-resuming version of hold.

holdFor

Arguments

:: Serialize a

=> Int
number of steps to hold the last emitted value for

-> Interval m (Blip a) a

Like hold, but it only "holds" the last emitted value for the given number of steps.

>>> let a        = holdFor 2 . inB 3 . count
>>> let (res, _) = stepAutoN' 7 a ()
>>> res
[Nothing, Nothing, Just 3, Just 4, Nothing, Nothing, Nothing]

holdFor_

Arguments

:: Int
number of steps to hold the last emitted value for

-> Interval m (Blip a) a

The non-serializing/non-resuming version of holdFor.

Composition with Interval

during :: Monad m => Auto m a b -> Auto m (Maybe a) (Maybe b)

Lifts an Auto m a b (transforming as into bs) into an Auto m (Maybe a) (Maybe b) (or, Interval m (Maybe a) b, transforming intervals of as into intervals of b.

It does this by "running" the given Auto whenever it receives a Just value, and skipping/pausing it whenever it receives a Nothing value.

>>> let a1        = during (sumFrom 0) . onFor 2 . pure 1
>>> let (res1, _) = stepAutoN' 5 a1 ()
>>> res1
[Just 1, Just 2, Nothing, Nothing, Nothing]

>>> let a2       = during (sumFrom 0) . offFor 2 . pure 1
>>> let (res2, _) = stepAutoN' 5 a2 ()
>>> res2
[Nothing, Nothing, Just 1, Just 2, Just 3]

(Remember that pure x is the Auto that ignores its input and constantly just pumps out x at every step)

Note the difference between putting the sumFrom "after" the offFor in the chain with during (like the previous example) and putting the sumFrom "before":

>>> let a3        = offFor 2 . sumFrom 0 . pure 1
>>> let (res3, _) = stepAutoN' 5 a3 ()
>>> res3
[Nothing, Nothing, Just 3, Just 4, Just 5]

In the first case (with a2), the output of pure 1 was suppressed by offFor, and during (sumFrom 0) was only summing on the times that the 1's were "allowed through"...so it only "starts counting" on the third step.

In the second case (with a3), the output of the pure 1 is never suppressed, and went straight into the sumFrom 0. sumFrom is always summing, the entire time. The final output of that sumFrom 0 is suppressed at the end with offFor 2.

compI :: Monad m => Interval m b c -> Interval m a b -> Interval m a c infixr 1

Composes two Intervals, the same way that . composes two Autos:

    (.)   :: Auto     m b c -> Auto     m a b -> Auto     m a c
    compI :: Interval m b c -> Interval m a b -> Interval m a c

Basically, if any Interval in the chain is "off", then the entire rest of the chain is "skipped", short-circuiting a la Maybe.

(Users of libraries with built-in inhibition semantics like Yampa and netwire might recognize this as the "default" composition in those other libraries)

As a contrived example, how about an Auto that only allows values through during a window...between, say, the second and fourth steps:

>>> let window start finish = onFor finish `compI` offFor start
>>> let a        = window 1 4 . count
>>> let (res, _) = stepAutoN' 5 a ()
>>> res
[Nothing, Just 2, Just 3, Just 4, Nothing, Nothing]

(Remember that count is the Auto that ignores its input and displays the current step count, starting with 1)

bindI :: Monad m => Interval m a b -> Auto m (Maybe a) (Maybe b)

Lifts (more technically, "binds") an Interval m a b into an Auto m (Maybe a) (Maybe b) (or an Interval m (Maybe a) b)

The given Auto is "run" only on the Just inputs, and paused on Nothing inputs.

It's kind of like during, but the resulting Maybe (Maybe b)) is "joined" back into a Maybe b.

bindI a == fmap join (during a)

This is really an alternative formulation of compI; typically, you will be using compI more often, but this form can also be useful (and slightly more general). Note that:

bindI f == compI f id

This combinator allows you to properly "chain" ("bind") together series of inhibiting Autos. If you have an Interval m a b and an Interval m b c, you can chain them into an Interval m a c.

    f             :: Interval m a b
    g             :: Interval m b c
    bindI g . f :: Interval m a c

(Users of libraries with built-in inhibition semantics like Yampa and netwire might recognize this as the "default" composition in those other libraries)

See compI for more examples of this use case.