Control.Auto.Process

Description

Various Autos for miscellaneous common processes.

Note that all of these can be turned into an equivalent version acting on Blip streams, with perBlip:

sumFrom         :: (Serialize a, Num a) => a -> Auto m a a
perBlip sumFrom :: (Serialize a, Num a) => a -> Auto m (Blip a) (Blip a)

Numerical

sumFrom

Arguments

:: (Serialize a, Num a)

=> a
initial count

-> Auto m a a

Outputs the running sum of all items passed so far, starting with an initial count.

The first output is the sum of the first input with the initial count. See the documentation for sumFromD for more information.

sumFrom x0 = mkAccum (+) x0

sumFrom_

Arguments

:: Num a

=> a
initial count

-> Auto m a a

The non-resuming/non-serializing version of sumFrom.

sumFromD

Arguments

:: (Serialize a, Num a)

=> a
initial count

-> Auto m a a

Like sumFrom, except the first output is the starting count.

>>> let a = sumFromD 5
>>> let Output y1 a' = stepAuto' a 10
>>> y1
5
>>> let Output y2 _  = stepAuto' a' 3
>>> y2
10

It's sumFrom, but "delayed".

Useful for recursive bindings, where you need at least one value to be able to produce its "first output" without depending on anything else.

sumFromD x0 = delay x0 . sumFrom x0

sumFromD_

Arguments

:: Num a

=> a
initial count

-> Auto m a a

The non-resuming/non-serializing version of sumFromD.

deltas :: (Serialize a, Num a) => Interval m a a

Returns the difference between the received input and the previous input. The first result is Nothing; if you have something you want the first result to be, you can use <|!> from Control.Auto.Interval, or just fromMaybe or maybe from Data.Maybe.

>>> let a = deltas
>>> let Output y1 a'  = stepAuto' a 5
>>> y1
Nothing
>>> let Output y2 a'' = stepAuto' a' 7
>>> y2
Just 2
>>> let Output y3 _   = stepAuto' a'' 4
>>> y3
Just (-3)

Usage with <|!>:

>>> let a = deltas <|!> pure 100
>>> let (ys, _) = overList' a [5,7,4]
>>> ys
[100, 2, -3]

Usage with fromMaybe:

>>> let a = fromMaybe 100 <$> deltas
>>> let (ys, _) = overList' a [5,7,4]
>>> ys
[100, 2, -3]

deltas_ :: Num a => Interval m a a

The non-resuming/non-serializing version of deltas.

Monoidal/Semigroup

mappender :: (Serialize a, Monoid a) => Auto m a a

Accumulates the monoid sum of all inputs, starting with mempty.

>>> let a = mappender
>>> let Output y1 a'  = stepAuto' a (Last (Just 4))
>>> y1
Last (Just 4)
>>> let Output y2 a'' = stepAuto' a' (Last Nothing)
>>> y2
Last (Just 4)
>>> let Output y3 _   = stepAuto' a'' (Last (Just 2))
>>> y3
Last (Just 2)

mappender = mkAccum mappend mempty

mappender_ :: Monoid a => Auto m a a

The non-resuming/non-serializing version of mappender.

mappendFrom :: (Serialize a, Semigroup a) => a -> Auto m a a

Accumulates the semigroup sum of all of its inputs; the same as mappender, but you have to provide a "starting" value.

>>> let a = mappendFrom (Max 0)
>>> let Output y1 a' = stepAuto' a (Max 3)
>>> y1
Max 3
>>> let Output y2 _  = stepAuto' a' (Max (-2))
>>> y2
Max 3

mappendFrom m0 = mkAccum (<>) m0

mappendFrom_ :: Semigroup a => a -> Auto m a a

The non-resuming/non-serializing version of mappender.