Control.Auto.Time

Description

This module contains various Auto transformers for manipulating the flow of time/stepping rate of an Auto.

Many of these are Auto "transformers", meaning that they take in an Auto and return a transformed Auto, with new stepping behavior.

For example, there is accelerate:

accelerate :: Monad m => Int -> Auto m a b -> Auto m a [b]

accelerate n turns an Auto into an Auto that "steps itself" n times for every single input/step. The result is a list of the results of each single step.

There are also various Autos for observing the passage of time (count) and actiong as a "delay" or a way to access the previously stepped values of an Auto.

TODO: dropping first few elements

A counter

count :: (Serialize b, Num b) => Auto m a b

A simple Auto that outputs the step count. First output is 0.

TODO: should be 1?

count_ :: Num b => Auto m a b

A non-resuming/non-serializing version of count.

Manipulating time

Delaying

lastVal

Arguments

:: Serialize a

=> a
initial value

-> Auto m a a

An Auto that returns the last value received by it. Given an "initial value" to output first.

This is (potentially) a very dangerous Auto, because its usage and its very existence opens the door to breaking denotative/declarative style and devolving into imperative style coding. However, when used where it is supposed to be used, it is more or less invaluable, and will be an essential part of many programs.

Its main usage is for dealing with bindings. If you ever are laying out recursive bindings in a high-level/denotative way, you need to have at least one value be able to have a "initial output" without depending on anything else. lastVal and delay allow you to do this.

See the recursive example for more information on the appropriate usage of lastVal and delay.

lastVal_

Arguments

:: a
initial value

-> Auto m a a

The non-resuming/non-serializing version of lastVal.

delay

Arguments

:: Serialize a

=> a
initial value

-> Auto m a a

An alias for lastVal; used in contexts where "delay" is more a meaningful description than "last value". All of the warnings for lastVal still apply, so you should probably read it if you haven't :)

delay_

Arguments

:: a
initial value

-> Auto m a a

The non-resuming/non-serializing version of delay.

delayList :: (Serialize a, Monad m) => [a] -> Auto m a a

Like delay, except has as many "initial values" as the input list. Outputs every item in the input list in order before returning the first received value.

delayList [y0] = delay y0

delayList_ :: Monad m => [a] -> Auto m a a

The non-resuming/non-serializing version of delayList.

delayN :: (Serialize a, Monad m) => Int -> a -> Auto m a a

Like delay, except delays the desired number of steps with the same initial output value.

delayN n x0 = delayList (replicate n x0)

delayN 1 x0 = delay x0

delayN_ :: Monad m => Int -> a -> Auto m a a

The non-resuming/non-serializing version of delayN

Stretching

stretch

Arguments

:: (Serialize b, Monad m)

=> Int
stretching factor

-> Auto m a b
Auto to stretch

-> Auto m a b

"stretch" an Auto out, slowing time. stretch n a will take one input, repeat the same output n times (ignoring input), and then take another.

>>> let a       = stretch 2 (mkAccum (+) 0)
>>> let (ys, _) = overList' a [1,8,5,4,3,7,2,0]
>>> ys
   [1,1,6,6,9,9,11,11]
-- [1,_,5,_,3,_,2 ,_ ] <-- the inputs

stretch_

Arguments

:: Monad m

=> Int
stretching factor

-> Auto m a b
Auto to stretch

-> Auto m a b

The non-resuming/non-serializing version of stretch.

stretchB

Arguments

:: Monad m

=> Int
stretching factor

-> Auto m a b
Auto to stretch

-> Auto m a (Blip b)

Like stretch, but instead of holding the the "stretched" outputs, emits a Blip every time the stretched Auto "progresses" (every n ticks)

See stretch for more information.

>>> let a = stretchB 2 (mkAccum (+) 0)
>>> let (ys, _) = overList' a [1,8,5,4,3,7,2,0]
>>> ys
[Blip 1, NoBlip, Blip 6, NoBlip, Blip 9, NoBlip, Blip 11, NoBlip]

Accelerating

accelerate

Arguments

:: Monad m

=> Int
acceleration factor

-> Auto m a b
Auto to accelerate

-> Auto m a [b]

accelerate n a turns an Auto a into an "accelerated" Auto, where every input is fed into the Auto n times. All of the results are collected in the output.

The same input is fed repeatedly n times.

>>> let a = accelerate 3 (mkAccum (+) 0)
>>> let Output ys1 a' = stepAuto' a 5
>>> ys1
[5,10,15]
>>> let Output ys2 _  = stepAuto' a' (-8)
>>> ys2
[7,-1,-9]

accelerateWith

Arguments

:: Monad m

=> a
default input value, during acceleration periods

-> Int
acceleration factor

-> Auto m a b
Auto to accelerate

-> Auto m a [b]

accelerateWith xd n a behaves like accelerate n a, except instead of repeating the same intput n times, that input is fed once; the rest of the n times, a is fed xd, the "default" x.

>>> let a = accelerateWith (-1) 3 (mkAccum (+) 0)
>>> let Output ys1 a' = stepAuto' a 5
>>> ys1
[5,4,3]    -- fed 5, then (-1) twice
>>> let Output ys2 _  = stepAuto' a' 14
>>> ys2
[17,16,15] -- fed 14, then (-1) twice

accelOverList

Arguments

:: Monad m

=> Auto m a b
Auto to accelerate

-> Auto m [a] [b]

Turns an Auto into an "accelerated" Auto; an Auto that takes an a and returns a b turns into an Auto that takes a list of a and returns a list of b from running that Auto on all of those a in order.

>>> let a = accelOverList (mkAccum (+) 0)
>>> let Output ys1 a' = stepAuto' a [3,9,2]
>>> ys1
[3, 12, 14]
>>> let Output ys2 _  = stepAuto' a' [8,5]
>>> ys2
[22, 30]

Skipping

TODO: say something

skipTo

Arguments

:: Monad m

=> a
default input value, during skipping periods

-> Auto m a (b, Blip c)
Auto to skip over, until each Blip.

-> Auto m a ([b], c)

Takes an Auto that produces (b, Blip c), and turns it into an Auto that produces ([b], c).

Basically, the new Auto "squishes together" the periods of output between each Blip occurrnece. All outputs between each Blip occurrence are accumulated and returned in the resulting [b].

It "does this" in the same manner as accelerateWith and fastForward: first feed the input, then step repeatedly with the default input value.

>>> let a :: Auto' Int (Int, Blip String)
        a = proc i -> do
                sums <- mkAccum (+) 0 -< i
                blp  <- inB 3         -< i     -- Blip at every 3 ticks.
                id    -< (sums, show <$> blp)
>>> let skipA :: Auto' Int ([Int], String)
        skipA = skipTo (-1) a
>>> let Output res1 skipA' = stepAuto' skipA 8
>>> res1
([8,7,6], "6")     -- fed 8 first, then (-1) repeatedly
>>> let Output res2 _      = stepAuto' skipA' 5
>>> res2
([11,10,9], "9")   -- fed 5 first, then (-1) repeatedly

fastForward

Arguments

:: Monad m

=> a
default input

-> Interval m a b
Auto to fastforward (past each Nothing)

-> Auto m a b

Takes an Auto and basically "removes" the periods in which it returns Nothing, by re-running the Auto with the same input until it outputs a Just.

Basically, it magically turns an Auto m a (Maybe b) into an Auto m a b, by skipping over each Nothing.

It does this in the same manner as accelerateWith and skipTo: first feed in the input, then step repeatedly with the default input value.

>>> let a :: Auto' Int (Maybe Int)
        a = proc i -> do
                sums <- mkAccum (+) 0 -< i
                id    -< if (i `mod` 5) == 0
                           then Just i
                           else Nothing
>>> let ffA :: Auto' Int Int
        ffA = fastForward (-1) a
>>> let Output y1 ffA' = stepAuto' ffA 7
>>> y1
5          -- went from 7 (Nothing), to 6 (Nothing), to 5 (Just 5)
>>> let Output y2 _    = stepAuto' ffA' (-9)
-5         -- went from -4 (Nothing) to -5 (Just (-5))

fastForwardEither

Arguments

:: Monad m

=> a
default input

-> Auto m a (Either c b)
Auto to fast-forward (past each Left)

-> Auto m a (b, [c])

Same behavior as fastForward, except accumulates all of the Left c outputs in a list.