Control.Auto.Effects

Description

This module exports the preferred ways of interacting with the underlying Monad of the Auto type, including accessing, executing, and manipulating such effects.

Running effects

Continually

arrM

Arguments

:: (a -> m b)
monadic function

-> Auto m a b

Lifts a "monadic function" (Kleisli arrow) into an Auto. Like arr, but with monadic functions. Composition of such Autos works just like the composition of Kleisli arrows, with '(<=<)':

arrM f . arrM g == arrM (f <=< g)

effect

Arguments

:: m b
monadic action to contually execute.

-> Auto m a b

Executes the given monadic action at every tick/step of the Auto, outputting the result. Particularly useful with Reader as the underlying Monad.

>>> let a = effect ask     :: Auto (Reader b) a b
>>> let r = evalAuto a ()  :: Reader b b
>>> runReader r "hello"
"hello"
>>> runReader r 123
123

Basically like pure, but instead of `pure :: b -> Auto m a b`, you have `effect :: m b -> Auto m a b`.

exec

Arguments

:: Monad m

=> m b
monadic action to contually execute.

-> Auto m a a

Executes the given monadic action at every tick/step of the Auto, but ignores the result. Basically acts like the id Auto, with attached effects.

From inputs

effects :: Monad m => Auto m (m a) a

Gets an input stream of monadic actions, executes them, and returns an output stream of their values.

On Blips

arrMB :: Monad m => (a -> m b) -> Auto m (Blip a) (Blip b)

Applies the "monadic function" to the contents of every incoming Blip, and replaces the contents of the Blip with the result of the function.

effectB :: Monad m => m b -> Auto m (Blip a) (Blip b)

Executes a monadic action with every incoming Blip, and replaces the contents of the Blip with the result of the action.

execB :: Monad m => m b -> Auto m (Blip a) (Blip a)

Executes a monadic action with every incoming Blip, and passes back that same Blip unchanged.

One-time effects

cache

Arguments

:: (Serialize b, Monad m)

=> m b
monadic action to execute and use the result of

-> Auto m a b

Executes a monadic action once (and only once), and continues to echo the result forevermore.

Like execOnce, exept outputs the result of the action.

Useful for loading resources in IO on the "first step", like a dictionary:

dictionary :: Auto IO a [String]
dictionary = cache (lines $ readFile "dictionary.txt")

execOnce

Arguments

:: Monad m

=> m b
monadic action to execute; result discared

-> Auto m a ()

Executes the monadic action once, and outputs '()' forevermore.

Pretty much like cache, but always outputs '()'.

cache_

Arguments

:: Monad m

=> m b
monadic action to execute and use the result of

-> Auto m a b

The non-resumable/non-serializable version of cache. Every time the Auto is deserialized/reloaded, it re-executes the action to retrieve the result again.

Useful in cases where you want to "re-load" an expensive resource on every startup, instead of saving it to in the save states.

dictionary :: Auto IO a [String]
dictionary = cache_ (lines $ readFile "dictionary.txt")

execOnce_

Arguments

:: Monad m

=> m b
monadic action to execute; result discared

-> Auto m a ()

The non-resumable/non-serializable version of execOnce. Every time the Auto is deserialized/reloaded, the action is re-executed again.

"Sealing off" monadic Autos

sealState :: (Monad m, Serialize s) => Auto (StateT s m) a b -> s -> Auto m a b

sealState_ :: Monad m => Auto (StateT s m) a b -> s -> Auto m a b

sealReader :: (Monad m, Serialize r) => Auto (ReaderT r m) a b -> r -> Auto m a b

sealReader_ :: Monad m => Auto (ReaderT r m) a b -> r -> Auto m a b