Control.Auto.Process

Description

Various Autos describing relationships following common processes, like sumFrom, whose output is the cumulative sum of the input.

Note that all of these can be turned into an equivalent version acting on Blip streams, with perBlip:

sumFrom n           :: Num a => Auto m a a
perBlip (sumFrom n) :: Num a => Auto m (Blip a) (Blip a)

Numerical

sumFrom

Arguments

:: (Serialize a, Num a)

=> a
initial count

-> Auto m a a

The stream of outputs is the cumulative/running sum of the inputs so far, starting with an initial count.

The first output takes into account the first input. See sumFromD for a version where the first output is the initial count itself.

sumFrom x0 = accum (+) x0

sumFrom_

Arguments

:: Num a

=> a
initial count

-> Auto m a a

The non-resuming/non-serializing version of sumFrom.

sumFromD

Arguments

:: (Serialize a, Num a)

=> a
initial count

-> Auto m a a

Like sumFrom, except the first output is the starting count.

>>> let a = sumFromD 5
>>> let Output y1 a' = stepAuto' a 10
>>> y1
5
>>> let Output y2 _  = stepAuto' a' 3
>>> y2
10

It's sumFrom, but "delayed".

Useful for recursive bindings, where you need at least one value to be able to produce its "first output" without depending on anything else.

sumFromD x0 = delay x0 . sumFrom x0

sumFromD_

Arguments

:: Num a

=> a
initial count

-> Auto m a a

The non-resuming/non-serializing version of sumFromD.

productFrom

Arguments

:: (Serialize a, Num a)

=> a
initial product

-> Auto m a a

The output is the running/cumulative product of all of the inputs so far, starting from an initial product.

productFrom x0 = accum (*) x0

productFrom_

Arguments

:: Num a

=> a
initial product

-> Auto m a a

The non-resuming/non-serializing version of productFrom.

deltas :: (Serialize a, Num a) => Interval m a a

The output is the the difference between the input and the previously received input.

First result is a Nothing, so you can use <|!> or fromInterval or fromMaybe to get a "default first value".

>>> streamAuto' deltas [1,6,3,5,8]
>>> [Nothing, Just 5, Just (-3), Just 2, Just 3]

Usage with <|!>:

>>> let a = deltas <|!> pure 100
>>> streamAuto' (deltas <|!> pure 100) [1,6,3,5,8]
[100, 5, -3, 2, 3]

Usage with fromMaybe:

>>> streamAuto' (fromMaybe 100 <$> deltas) [1,6,3,5,8]
[100, 5, -3, 2, 3]

deltas_ :: Num a => Interval m a a

The non-resuming/non-serializing version of deltas.

Monoidal/Semigroup

mappender :: (Serialize a, Monoid a) => Auto m a a

The output is the running/cumulative mconcat of all of the input seen so far, starting with mempty.

>>> streamauto' mappender . map Last $ [Just 4, Nothing, Just 2, Just 3]
[Last (Just 4), Last (Just 4), Last (Just 2), Last (Just 3)]
>>> streamAuto' mappender ["hello","world","good","bye"]
["hello","helloworld","helloworldgood","helloworldgoodbye"]

mappender = accum mappend mempty

mappender_ :: Monoid a => Auto m a a

The non-resuming/non-serializing version of mappender.

mappendFrom

Arguments

:: (Serialize a, Semigroup a)

=> a
initial value

-> Auto m a a

The output is the running <>-sum (mappend for Semigroup) of all of the input values so far, starting with a given starting value. Basically like mappender, but with a starting value.

>>> streamAuto' (mappendFrom (Max 0)) [Max 4, Max (-2), Max 3, Max 10]
[Max 4, Max 4, Max 4, Max 10]

mappendFrom m0 = accum (<>) m0

mappendFrom_

Arguments

:: Semigroup a

=> a
initial value

-> Auto m a a

The non-resuming/non-serializing version of mappender.